Buffer

Buffer

Jusdl.Utilities.Buffer — Type.

Buffer{T, N} <: AbstractBuffer{T, N}

N dimensional Buffer with element type T.

Buffer(::Type{T}, shape::Int...; [mode::Symbol, [callbacks::Vector{Callback}, [name;;String]]])

Constructs a Buffer of shape shape and element type T. callbacks is the vector of additional callbacks for event monitoring capability and name is the name of Buffer. mode is the mode of the buffer with the following properties:

normal: Data can be written into buffer until the buffer is full. When the buffer is full, no more writing is possible. When read, the last written element is returned without deleting the returned element.

cyclic: Data can be written into buffer until the buffer is full. When the buffer is full, data is written into the the buffer after shifting the buffer data to right. When read, the last written element is returned without deleting the returned element.

lifo: Data can be written into buffer until the buffer is full. When the buffer is full, no more writing is possible. When read, last element written to the buffer is returned by deleting the returned element.

fifo: Data can be written into buffer until the buffer is full. When the buffer is full, no more writing is possible. When read, fist element written to the buffer is returned by deleting the returned element.

Buffer(shape::Int..., [mode::Symbol, [callbacks::Vector{Callback}, [name::String]]])

Constructs a Buffer of shapeshape`