Printer

Basic Operation of Printers

See Basic Operation of Writers since the operation of Writer and that of Printer is very similar.

Full API

Causal.Printer — Type

Printer(input=Inport(); buflen=64, plugin=nothing, callbacks=nothing, name=Symbol()) where T Constructs a Printer with input bus input. buflen is the length of its internal buflen. plugin is data proccessing tool.

Base.print — Method

print(printer::Printer, td, xd)

Prints xd corresponding to xd to the console.

Base.open — Method

open(printer::Printer)

Does nothing. Just a common interface function ot AbstractSink interface.

Base.close — Method

close(printer::Printer)

Does nothing. Just a common interface function ot AbstractSink interface.