Generators

FunctionGenerator

`FunctionGenerator is the generic function generator. The output waveform is shaped by its output function. A FunctionGenerator can be construted by specifying its output function outputfunc.

Jusdl.Components.Sources.FunctionGenerator — Type

FunctionGenerator(outputfunc)

Constructs a FunctionGenerator with the output of the form x(t) = f(t) where $f$ is outputfunc.

Basic Operation AbstractSource

AnAbstractSource is a subtype of AbstractComponent. (See Components for more information.) An AbstractComponent has input and output for data flow. The AbstractComponent reads data from input and writes data to output. Since the input-output relation of AbstractSource depends on just the current time t, Sources does not have inputs since they do not read input values. They just need time t to compute its output. During their evolution, the AbstractComponent read time t from their trigger links, computes their output according to their output function and writes its computed output to their output busses. An AbstractComponent also writes true to their handshake links in signal that the evolution is succeeded. To further clarify the operation of AbstractSource, let us do some examples.

julia> using Jusdl # hide

julia> outputfunc(t) = t * exp(t) + sin(t)
outputfunc (generic function with 1 method)

julia> gen = FunctionGenerator(outputfunc)
FunctionGenerator(outputfunc:outputfunc, nout:1)

We constructed a FunctionGenerator which is an AbstractSource.

julia> gen isa AbstractSource
true

Full API

Jusdl.Components.Sources.SinewaveGenerator — Type

SinewaveGenerator(;amplitude=1., frequency=1., phase=0., delay=0., offset=0.)

Constructs a SinewaveGenerator with output of the form

\[ x(t) = A sin(2 \pi f (t - \tau) + \phi) + B\]

where $A$ is amplitude, $f$ is frequency, $\tau$ is delay and $\phi$ is phase and $B$ is offset.

Jusdl.Components.Sources.DampedSinewaveGenerator — Type

DampedSinewaveGenerator(;amplitude=1., decay=-0.5, frequency=1., phase=0., delay=0., offset=0.)

Constructs a DampedSinewaveGenerator which generates outputs of the form

\[ x(t) = A e^{\alpha t} * sin(2 \pi f (t - \tau) + \phi) + B\]

where $A$ is amplitude, $\alpha$ is decay, $f$ is frequency, $\phi$ is phase, $\tau$ is delay and $B$ is offset.

Jusdl.Components.Sources.SquarewaveGenerator — Type

SquarewaveGenerator(;level1=1., level2=0., period=1., duty=0.5, delay=0.)

Constructs a SquarewaveGenerator with output of the form

\[ x(t) = \left\{\begin{array}{lr} A_1 + B, & kT + \tau \leq t \leq (k + \alpha) T + \tau \\ A_2 + B, & (k + \alpha) T + \tau \leq t \leq (k + 1) T + \tau \end{array} \right. \quad k \in Z\]

where $A_1$, $A_2$ is level1 and level2, $T$ is period, $\tau$ is delay $\alpha$ is duty.

Jusdl.Components.Sources.TriangularwaveGenerator — Type

TriangularwaveGenerator(;amplitude=1, period=1, duty=0.5, delay=0, offset=0)

Constructs a TriangularwaveGenerator with output of the form

\[ x(t) = \left\{\begin{array}{lr} \dfrac{A t}{\alpha T} + B, & kT + \tau \leq t \leq (k + \alpha) T + \tau \\ \dfrac{A (T - t)}{T (1 - \alpha)} + B, & (k + \alpha) T + \tau \leq t \leq (k + 1) T + \tau \end{array} \right. \quad k \in Z\]

where $A$ is amplitude, $T$ is period, $\tau$ is delay $\alpha$ is duty.

Jusdl.Components.Sources.ConstantGenerator — Type

ConstantGenerator(;amplitude=1.)

Constructs a Constructs with output of the form

\[ x(t) = A\]

where $A$ is `amplitude.

Jusdl.Components.Sources.RampGenerator — Type

RampGenerator(;scale=1)

Constructs a RampGenerator with output of the form

\[ x(t) = \alpha t\]

where $\alpha$ is the scale.

Jusdl.Components.Sources.StepGenerator — Type

StepGenerator(;amplitude=1, delay=0, offset=0)

Constructs a StepGenerator with output of the form

\[ x(t) = \left\{\begin{array}{lr} B, & t \leq 0 \\ A + B, & t > 0 \end{array} \right.\]

Jusdl.Components.Sources.ExponentialGenerator — Type

ExponentialGenerator(;scale=1, decay=-1)

Constructs an ExponentialGenerator with output of the form

\[ x(t) = A e^{\alpha t}\]

where $A$ is scale, $\alpha$ is decay.

Jusdl.Components.Sources.DampedExponentialGenerator — Type

DampedExponentialGenerator(;scale=1, decay=-1)

Constructs an DampedExponentialGenerator with outpsuts of the form

\[ x(t) = A t e^{\alpha t}\]

where $A$ is scale, $\alpha$ is decay.