DAESystem

Basic Operation of DAESystem

The basic operation of DAESystem is the same as Basic Operation of ODESystem.

Full API

Jusdl.Components.Systems.DynamicSystems.DAESystem — Type

DAESystem(input, output, statefunc, outputfunc, state, stateder, t, diffvars; solver=DAESolver)

Construsts a DAESystem with input and output. statefunc is the state function and outputfunc is the output function. DAESystem is represented by the following equations.

\[ \begin{array}{l} 0 = f(out, dx, x, u, t) \\ y = f(x, u, t) \end{array}\]

where $t$ is the time t, $x$ is state, $dx$ is the value of the derivative of the state stateder, $u$ is the value of input and $y$ is the value of output at time $t$. solver is used to solve the above differential equation.

The signature of statefunc must be of the form

function statefunc(out, dx, x, u, t)
    out .= ... # Update out
emd

and the signature of outputfunc must be of the form

function outputfunc(x, u, t)
    y = ... # Compute y 
    return y
end