DDESystem

Construction of DDESystem

A DDESystem is represented by the following state equation

\[ \dot{x} = f(x, h, u, t) \quad t \geq t_0\]

where $t$ is the time, $x$ is the value of the state, $u$ is the value of the input. $h$ is the history function for which

\[ x(t) = h(t) \quad t \leq t_0\]

and by the output equation

\[ y = g(x, u, t) \]

where $y$ is the value of the output.

As an example, consider a system with the state equation

\[ \begin{array}{l} \dot{x} = -x(t - \tau) \quad t \geq 0 \\ x(t) = 1. -\tau \leq t \leq 0 \\ \end{array}\]

First, we define the history function histfunc,

julia> const out = zeros(1)
1-element Array{Float64,1}:
 0.0

julia> histfunc(out, u, t) = (out .= 1.);

Note that histfunc mutates a vector out. This mutation is for performance reasons. Next the state function can be defined

julia> function statefunc(dx, x, h, u, t)
           h(out, u, t - tau) # Update out vector
           dx[1] = out[1] + x[1]
       end
statefunc (generic function with 1 method)

and let us take all the state variables as outputs. Thus, the output function is

julia> outputfunc(x, u, t) = x
outputfunc (generic function with 1 method)

Next, we need to define the history for the system. History is defined by specifying a history function, and the type of the lags. There may be two different lag: constant lags which are independent of the state variable $x$ and the dependent lags which are mainly the functions of the state variable $x$. Note that for this example, the have constant lags. Thus,

julia> tau = 1
1

julia> conslags = [tau]
1-element Array{Int64,1}:
 1

At this point, we are ready to construct the system.

julia> ds = DDESystem((statefunc, histfunc), outputfunc, [1.],  0., nothing, Outport())
ERROR: MethodError: no method matching DDESystem(::Tuple{typeof(Main.ex-dde_system_ex.statefunc),typeof(Main.ex-dde_system_ex.histfunc)}, ::typeof(Main.ex-dde_system_ex.outputfunc), ::Array{Float64,1}, ::Float64, ::Nothing, ::Outport{Outpin{Float64}})
Closest candidates are:
  DDESystem(::CL, ::DL, ::RH, ::HST, ::RO, ::ST, !Matched::IP, !Matched::OP, !Matched::TR, !Matched::HS, !Matched::CB, !Matched::Symbol, !Matched::ID, !Matched::Float64, !Matched::MA, !Matched::MK, !Matched::SA, !Matched::SK, !Matched::AL, !Matched::IT) where {CL, DL, RH, HST, RO, ST, IP, OP, TR, HS, CB, ID, MA, MK, SA, SK, AL, IT} at /home/sari/.julia/dev/Jusdl/src/components/systems/dynamicalsystems/ddesystems.jl:45

Basic Operation of DDESystem

The basis operaiton of DDESystem is the same as those of other dynamical systems. When triggered from its trigger link, the DDESystem reads its time from its trigger link, reads input, solves its differential equation, computes its output and writes the computed output to its output bus. To drive DDESystem, we must first launch it,

julia> iport, trg, hnd = Inport(), Outpin(), Inpin{Bool}()
(Inport(numpins:1, eltype:Inpin{Float64}), Outpin(eltype:Float64, isbound:false), Inpin(eltype:Bool, isbound:false))

julia> connect!(ds.output, iport)
ERROR: UndefVarError: ds not defined

julia> connect!(trg, ds.trigger)
ERROR: UndefVarError: ds not defined

julia> connect!(ds.handshake, hnd)
ERROR: UndefVarError: ds not defined

julia> task = launch(ds)
ERROR: UndefVarError: ds not defined

julia> task2 = @async while true
           all(take!(iport) .=== NaN) && break
           end
Task (failed) @0x00007f959c5b4760
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:97 [inlined]
_broadcast_getindex_evalf at ./broadcast.jl:631 [inlined]
_broadcast_getindex at ./broadcast.jl:604 [inlined]
getindex at ./broadcast.jl:564 [inlined]
macro expansion at ./broadcast.jl:910 [inlined]
macro expansion at ./simdloop.jl:77 [inlined]
copyto! at ./broadcast.jl:909 [inlined]
copyto! at ./broadcast.jl:864 [inlined]
copy at ./broadcast.jl:840 [inlined]
materialize at ./broadcast.jl:820 [inlined]
take!(::Inport{Inpin{Float64}}) at /home/sari/.julia/dev/Jusdl/src/connections/port.jl:178
macro expansion at ./none:2 [inlined]
(::Main.ex-dde_system_ex.var"#1#2")() at ./task.jl:358

When launched, ds is drivable. To drive ds, we can use the syntax drive(ds, t) or put!(ds.trigger, t) where t is the time until which ds is to be driven.

julia> put!(trg, 1.)

When driven, ds reads the time t from its trigger link, (since its input is nothing, ds does nothing during its input reading stage), solves its differential equation, computes output and writes the value of its output to its output bus. To signify, the step was taken with success, ds writes true to its handshake which must be read to further drive ds. For this, we can use the syntax approve!(ds) or take!(ds.handshake).

julia> take!(hnd)
ERROR: UndefRefError: access to undefined reference

We can continue to drive ds.

julia> for t in 2. : 10.
           put!(trg, t)
           take!(hnd)
       end
ERROR: UndefRefError: access to undefined reference

When launched, we constructed a task whose state is running which implies that ds can be driven.

julia> task
ERROR: UndefVarError: task not defined

julia> task2
Task (failed) @0x00007f959c5b4760
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:97 [inlined]
_broadcast_getindex_evalf at ./broadcast.jl:631 [inlined]
_broadcast_getindex at ./broadcast.jl:604 [inlined]
getindex at ./broadcast.jl:564 [inlined]
macro expansion at ./broadcast.jl:910 [inlined]
macro expansion at ./simdloop.jl:77 [inlined]
copyto! at ./broadcast.jl:909 [inlined]
copyto! at ./broadcast.jl:864 [inlined]
copy at ./broadcast.jl:840 [inlined]
materialize at ./broadcast.jl:820 [inlined]
take!(::Inport{Inpin{Float64}}) at /home/sari/.julia/dev/Jusdl/src/connections/port.jl:178
macro expansion at ./none:2 [inlined]
(::Main.ex-dde_system_ex.var"#1#2")() at ./task.jl:358

As long as the state of the task is running, ds can be driven. To terminate task safely, we need to terminate the ds.

julia> put!(trg, NaN)

Note that the state of task is done which implies that ds is not drivable any more.

Note that the output values of ds is written to its output bus.

julia> iport[1].link.buffer
ERROR: UndefRefError: access to undefined reference

Full API

Jusdl.@def_dde_system — Macro

@def_dde_system

Used to define DDE system

Jusdl.DDESystem — Type

DDESystem(; constantlags, depslags, righthandside, history, readout, state, input, output)

Construct a generic DDE system

Jusdl.DelayFeedbackSystem — Type

DDESystem(; constantlags, depslags, righthandside, history, readout, state, input, output)

Constructs DelayFeedbackSystem