DiscreteSystem

Construction of DiscreteSystem

DiscreteSystems evolve by the following discrete time difference equation.

\[ x_{k + 1} = f(x_k, u_k, k) \\ y_k = g(x_k, u_k, k)\]

where $x_k$ is the state, $y_k$ is the value of output and $u_k$ is the value of input at discrete time t. $f$ is the state function and $g$ is the output function of the system. See the main constructor.

Jusdl.Components.Systems.DynamicSystems.DiscreteSystem — Type

DiscreteSystem(input, output, statefunc, outputfunc, state, t;  solver=DiscreteSolver)

Construct a DiscreteSystem with input and output. statefunc is the state function and outputfunc is the output function of DiscreteSystem. state is the state and t is the current time of t. solver is numerical difference eqaution solver of the system. The system is represented by

\[ \begin{array}{l} x_{k + 1} = f(x_k, u_k, k) \\ y_k = g(x_k, u_k, k) \end{array}\]

where $x_k$ is the state, $y_k$ is the value of output, $u_k$ is the value of input at dicrete time $k$. $f` is `statefunc` and$g`isoutputfunc`.

The signature of statefunc must be of the form

function statefunc(dx, x, u, t)
    dx = ...   # Update dx
end

and the signature of outputfunc must be of the form

function outputfunc(x, u, t)
    y = ...   # Compute y
    return y
end

Example

julia> sfunc(dx,x,u,t) = (dx .= 0.5x)
sfunc (generic function with 1 method)

julia> ofunc(x, u, t) = x
ofunc (generic function with 1 method)

julia> ds = DiscreteSystem(Bus(1), Bus(1), sfunc, ofunc, [1.], 0.)
DiscreteSystem(state:[1.0], t:0.0, input:Bus(nlinks:1, eltype:Float64, isreadable:false, iswritable:false), output:Bus(nlinks:1, eltype:Float64, isreadable:false, iswritable:false))

Let us construct an example DiscreteSystem.