StaticSystems

Basic Operation of StaticSystems

A static system is a system whose output y at time t depends on the current time t and the value of its input u. The input-output relation of a static systems is represented by its output function outputfunc which is of the form

\[ y = g(u, t)\]

where g is the output function outputfunc. Note that outputfunc is expected to have two inputs, the value u of the input and the current time t. The simulation in Jusdl is a clocked-simulation, that is the data flowing through the input and output connections of components is actually sampled at time t. Therefore, for example, the system modeled by

\[y(t) = g(u(t),t)\]

is actually sampled at clock ticks t which is generated by a Clock. Therefore the sampled system corresponds to

\[y[k] = g(u_k, t_k)\]

where $k$ is $k_i T_s$ where $k_i$ is an integer number, $T_s$ is the sampling interval. $T_s$ corresponds to sampling time dt of Clock. Thus, the system given above is coded like

function g(u, t)
    # Define the relation `y = g(u, t)`
end

For further clarity, let us continue with a case study. Consider the following static system,

\[ y(t) = g(u(t), t) = \left[ \begin{array}{l} t u_1(t) \\ sin(u_1(t)) \\ cos(u_2(t)) \end{array} \right]\]

Note that the number of inputs is 2 and the number of outputs of is 3. To define such a system, the output function is written as

julia> g(u, t) = [t * u[1], sin(u[1]), cos(u[2])]
g (generic function with 1 method)

Note that the function g is defined in such a way that the input value u is sampled, which implies u is not a vector of function but is a vector of real. Having defined output function outputfunc, the system can be constructed.

julia> ss = StaticSystem(g, Inport(2), Outport(3))
ERROR: MethodError: no method matching StaticSystem(::typeof(Main.ex-static_system_ex.g), ::Inport{Inpin{Float64}}, ::Outport{Outpin{Float64}})
Closest candidates are:
  StaticSystem(::RO, ::IP, ::OP, !Matched::TR, !Matched::HS, !Matched::CB, !Matched::Symbol, !Matched::ID) where {RO, IP, OP, TR, HS, CB, ID} at /home/sari/.julia/dev/Jusdl/src/components/systems/staticsystems/staticsystems.jl:30

Note the construction of input bus Inport(2) and output bus Outport(3) by recalling that the number of input is 2 and the number of output is 3.

A StaticSystem evolves by being triggered through its trigger pin. When triggered from its trigger pin, a StaticSystem reads the current time t from its trigger pin and computes its output y according to its output function outputfunc and writes its output y(t) to its output port (if output port exists since output port may be nothing depending on the relation defined by outputfunc). When constructed, a StaticSystem is not ready to be triggered since its trigger pin is not writeable.

julia> ss.trigger.link
ERROR: UndefVarError: ss not defined

To make ss drivable, we need to construct the ports and pins for input-output and signaling.

julia> oport, iport, trg, hnd = Outport(length(ss.input)), Inport(length(ss.output)), Outpin(), Inpin{Bool}()
ERROR: UndefVarError: ss not defined

julia> connect!(oport, ss.input)
ERROR: UndefVarError: ss not defined

julia> connect!(ss.output, iport)
ERROR: UndefVarError: ss not defined

julia> connect!(trg, ss.trigger)
ERROR: UndefVarError: ss not defined

julia> connect!(ss.handshake, hnd)
ERROR: UndefVarError: ss not defined

julia> task = launch(ss)
ERROR: UndefVarError: ss not defined

julia> taskout = @async while true
           all(take!(iport) .=== NaN) && break
           end
Task (failed) @0x00007f959eab2980
UndefVarError: iport not defined
macro expansion at ./none:2 [inlined]
(::Main.ex-static_system_ex.var"#1#2")() at ./task.jl:358

Now, ss is drivable from its trg pin.

julia> ss.trigger.link
ERROR: UndefVarError: ss not defined

Now let us drive ss.

julia> put!(trg, 1.)
ERROR: UndefVarError: trg not defined

As this point ss wait for its to be written. Let us write some data to oport.

julia> put!(oport, [10., 10.])
ERROR: UndefVarError: oport not defined

ss read the value u of its input(since ss.input is connected to oport), read the current time t, and computed its output value y and wrote it its output port. To signal that it succeeded to be take the step, it put a true to its handshake which needs to be taken.

julia> hnd.link
ERROR: UndefVarError: hnd not defined

julia> take!(hnd)
ERROR: UndefVarError: hnd not defined

We can see the current data in the output of ss through iport (since iport is connected to ss.output)

julia> iport[1].link.buffer
ERROR: UndefVarError: iport not defined

Let us further drive ss.

julia> for t in 2. : 10.
           put!(trg, t)
           put!(oport, [10 * t, 20 * t])
           take!(hnd)
       end
ERROR: UndefVarError: trg not defined

The data written to the output of ss is also written to the internal buffers of output.

julia> iport[1].link.buffer
ERROR: UndefVarError: iport not defined

In addition to the generic StaticSystem, Jusdl provides some well-known static systems given in the next section.

Full API

Jusdl.@def_static_system — Macro

@def_static_system

Used to define static systems

Jusdl.StaticSystem — Type

StaticSystem(; readout, input, output)

Generaic static system.

Jusdl.Adder — Type

Adder(signs=(+,+))

Construts an Adder with input bus input and signs signs. signs is a tuplle of + and/or -. The output function g of Adder is of the form,

\[ y = g(u, t) = \sum_{j = 1}^n s_k u_k\]

where n is the length of the input, $s_k$ is the kth element of signs, $u_k$ is the kth value of input and $y$ is the value of output. The default value of signs is all +.

Example

julia> adder = Adder(signs=(+, +, -));

julia> adder.readout([3, 4, 5], 0.) == 3 + 4 - 5
true

Jusdl.Multiplier — Type

Multiplier(ops=(*,*))

Construts an Multiplier with input bus input and signs signs. signs is a tuplle of * and/or /. The output function g of Multiplier is of the form,

\[ y = g(u, t) = \prod_{j = 1}^n s_k u_k\]

where n is the length of the input, $s_k$ is the kth element of signs, $u_k$ is the kth value of input and $y$ is the value of the output. The default value of signs is all *.

Example

julia> mlt = Multiplier(ops=(*, *, /));

julia> mlt.readout([3, 4, 5], 0.) == 3 * 4 / 5
true

Jusdl.Gain — Type

Gain(input; gain=1.)

Constructs a Gain whose output function g is of the form

\[ y = g(u, t) = K u\]

where $K$ is gain, $u$ is the value of input and y is the value of output.

Example

julia> K = [1. 2.; 3. 4.];

julia> sfunc = Gain(input=Inport(2), gain=K);

julia> sfunc.readout([1., 2.], 0.) == K * [1., 2.]
true

Jusdl.Terminator — Type

Terminator(input::Inport)

Constructs a Terminator with input bus input. The output function g is eqaul to nothing. A Terminator is used just to sink the incomming data flowing from its input.

Jusdl.Memory — Type

Memory(delay=1.; initial::AbstractVector{T}=zeros(1), numtaps::Int=5, t0=0., dt=0.01, callbacks=nothing, 
    name=Symbol()) where T

Constructs a 'Memorywith input businput. A 'Memory delays the values of input by an amount of numdelay. initial determines the transient output from the Memory, that is, until the internal buffer of Memory is full, the values from initial is returned.

Example

julia> Memory(delay=0.1)
Memory(delay:0.1, numtaps:5, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

julia> Memory(delay=0.1, numtaps=5)
Memory(delay:0.1, numtaps:5, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

Jusdl.Coupler — Type

Coupler(conmat::AbstractMatrix, cplmat::AbstractMatrix)

Constructs a coupler from connection matrix conmat of size $n \times n$ and coupling matrix cplmat of size $d \times d$. The output function g of Coupler is of the form

\[ y = g(u, t) = (E \otimes P) u\]

where $\otimes$ is the Kronecker product, $E$ is conmat and $P$ is cplmat, $u$ is the value of input and y is the value of output.

Jusdl.Differentiator — Type

Differentiator(kd=1; callbacks=nothing, name=Symbol())

Consructs a Differentiator whose input output relation is of the form

\[ y(t) = k_d \dot{u}(t)\]

where $u(t)$ is the input and $y(t)$ is the output and $kd$ is the differentiation constant.