Busses

A Bus is actually is a bunch of links. Reading from and writing into data is performed as in the case of Links.

Construction of Bus

A Bus is constructed by specifying its element type T, number of links nlinks and the buffer length of its links.

Jusdl.Connections.Bus — Type

Bus{T}([nlinks::Int=1, [ln::Int=64]]) where T

Constructs a Bus consisting of nlinks links. ln is the buffer length and T is element type of the links.

Data Flow through Busses

Data flow through the Busses is very similar to the case in Links. See Data Flow through Links for information about data flow through Links. Runnable tasks must be bound to the links of the busses for data flow through the Bus. Again, put! and take! functions are used to write data from a Bus and read from data from a Bus.

Base.put! — Method

put!(bus::Bus, vals)

Puts vals to bus. Each item in vals is putted to the links of the bus.

Warning

The bus must be writable to be read. That is, there must be a runnable tasks bound to links of the bus that reads data from bus. See launch

Example

julia> b = Bus(2);

julia> t = launch(b);

julia> put!(b, [1., 2.])
┌ Info: Took 
└   val = 1.0
┌ Info: Took 
└   val = 2.0
2-element Array{Float64,1}:
 1.0
 2.0

Base.take! — Method

take!(bus::Bus)

Takes an element from bus. Each link of the bus is a read and a vector containing the results is returned.

Warning

The bus must be readable to be read. That is, there must be a runnable tasks bound to links of the bus that writes data to bus. See launch

Example

julia> b = Bus(2);

julia> t = launch(b, [[rand() for i = 1 : 5] for j = 1 : 2])
2-element Array{Task,1}:
 Task (runnable) @0x00007f9634734280
 Task (runnable) @0x00007f96347344f0

julia> take!(b)
2-element Array{Float64,1}:
 0.6216364091492494
 0.0781964275368685

Indexing and Iteration of Busses

Busses can be indexed similarly to the arrays in Julia. When indexed, corresponding link of the bus is returned.

julia> using Jusdl # hide

julia> b = Bus(3)
Bus(nlinks:3, eltype:Float64, isreadable:false, iswritable:false)

julia> b[1]
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[end]
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[:]
3-element Array{Link{Union{Missing, Float64}},1}:
 Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[1] = Link()
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[1:2] = [Link(), Link()]
2-element Array{Link{Union{Missing, Float64}},1}:
 Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)