Links

Links are built on top of Channels of Julia. They are used as communication primitives for Tasks of Julia. A Link includes a Channel and a Buffer. The mode of the buffer is Cyclic.(see Buffer Modes for information on buffer modes). Every item sent through a Link is sent through the channel of the Link and written to the Buffer so that all the data flowing through a Link is recorded. Any type of Julia can be transmitted through a Link, even if user-defined types.

A Link has a buffer to record flowing data and channel to transmit data between tasks. The Links can be connected to each other. To manage connection of the Links, Pin types are used. Thus, a Link has one pair of pin: leftpin and rightpin. When connected, a link has a master link and slaves links. Let us assume that links l1 and l2 are connected to each other and data flows from l1 to l2. Then, l1 is the master link of l2, similarly, l2 is the slave links of l1.

Construction of Links

The construction of a Link is very simple. See the main constructor.

Jusdl.Connections.Link — Type

Link{T}([ln::Int=64]) where T

Constructs a Link with element type T and buffer length ln. The buffer element type of T and mode is Cyclic.

That is, to construct a Link, just specify the its buffer length and the type of element to be flow through the Link. See the examples.

julia> using Jusdl # hide

julia> l1 = Link{Int}(5)   # A `Link` with buffer size of `5` and `Int` buffer element type.
Link(state:open, eltype:Union{Missing, Int64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> l2 = Link{Matrix{Float64}}(10)   # A `Link` with buffer size of `10` and `Matrix{Float64}` buffer element type.
Link(state:open, eltype:Union{Missing, Array{Float64,2}}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

Note that when initialized, a Link has no master and slaves. The Lin is not readable and writable since there is no active tasks bound ot Link.

Warning

Since the Link type is primarily used to transmit data or message between tasks in order there must be active tasks bound the Link.

Similar to the case of Buffers, the data type that can flow the Link can be any Julia type, even a user-defined type.

julia> struct Object
           x::Int
       end

julia> l = Link{Object}(3)     # A `Link` that with element type `Object` with buffer size `3`.
Link(state:open, eltype:Union{Missing, Main.##ex-#382.Object}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

Connection and Disconnection of Links

The Links can be connected to each other via connect function.

Jusdl.Connections.connect — Function

connect(master::Link, slave::Link)

Connects master to slave. When connected, the flow is from master to slave.

Example

julia> l1 = Link();

julia> l2 = Link();

julia> isconnected(l1, l2)
false

julia> connect(l1, l2)

julia> isconnected(l1, l2)
true

connect(master::AbstractVector{<:Link}, slave::AbstractVector{<:Link})

Connect master links to slave links by applying one-to-one matching of master links to slave links.

connect(links...)

Connect each link of links in the form of a path.

Example

julia> l1, l2, l3 = Link(), Link(), Link();

julia> isconnected(l1, l2)
false

julia> isconnected(l2, l3)
false

julia> connect(l1, l2, l3)

julia> isconnected(l1, l2)
true

julia> isconnected(l2, l3)
true

Similarly Links can be disconnected.

Jusdl.Connections.disconnect — Function

disconnect(link1::Link, link2::Link)

Disconnects link1 and link2. The order of arguments is not important.

Example

julia> ls = [Link() for i = 1 : 2];

julia> connect(ls[1], ls[2])

julia> disconnect(ls[1], ls[2])

julia> isconnected(ls[1], ls[2])
false

Warning

Note that the order or arguments is important when the links are connected. connect(l1, l2) connects l1 and l2 such that l1 drives l2, i.e., data flows from l1 to l2. In other words, l1 is the master link and l2 is the slave link. However, the order of arguments is not important when the links are disconnected. disconnect(l1, l2) does the same thing with disconnect(l2, l1), i.e., it justs breaks the connection between l2 and l1.

Full API

Base.put! — Function

put!(link::Link, val)

Puts val to link. val is handed over to the channel of link. val is also written in to the buffer of link.

Warning

link must be writable to put val. See launch

Base.take! — Function

take!(link::Link)

Take an element from link.

Warning

link must be readable to take value. See launch

Base.close — Function

close(link)

Closes link. All the task bound the link is also terminated safely.

Base.isopen — Function

open(link::Link)

Returns true if link is open. A link is open if its channel is open.

Base.isreadable — Function

isreadable(link::Link)

Returns true if link is readable. When link is readable, data can be read from link with take function.

Base.iswritable — Function

writable(link::Link)

Returns true if link is writable. When link is writable, data can be written into link with put function.

Jusdl.Connections.isfull — Function

isfull(link::Link)

Returns true if the buffer of link is full.

Jusdl.Connections.isconnected — Function

isconnected(link1, link2)

Returns true if link1 is connected to link2. The order of the arguments are not important.

Jusdl.Connections.hasslaves — Function

hasslaves(link::Link)

Returns true if link has slave links.

Jusdl.Connections.hasmaster — Function

hasmaster(link::Link)

Returns true if link has a master link.

Jusdl.Connections.getmaster — Function

getmaster(link::Link)

Returns the master of link.

Jusdl.Connections.getslaves — Function

getslaves(link::Link)

Returns the slaves of link.

Jusdl.Connections.snapshot — Function

snapshot(link::Link)

Returns all the data of the buffer of link.

Jusdl.Connections.UnconnectedLinkError — Type

UnconnectedLinkError <: Exception

Exception thrown when the links are not connected to each other.

Jusdl.Connections.Pin — Type

Pin()

Constructs a Pin. A Pin is the auxilary type to monitor connection status of Links. See Link

Jusdl.Connections.findflow — Function

findflow(link1::Link, link2::Link)

Returns a tuple of (masterlink, slavelink) where masterlink is the link that drives the other and slavelink is the link that is driven by the other.

Example

julia> ls = [Link() for i = 1 : 2];

julia> connect(ls[1], ls[2])

julia> findflow(ls[2], ls[1]) .== (ls[1], ls[2])
(true, true)

Jusdl.Connections.insert — Function

insert(master::Link, slave::Link, new::Link)

Inserts the new link between the master link and slave link. The master is connected to new, and new is connected to slave.

Example

julia> ls = [Link() for i = 1 : 3];  

julia> connect(ls[1], ls[2]) 

julia> insert(ls[1], ls[2], ls[3])

julia> isconnected(ls[1], ls[2])
false

julia> isconnected(ls[1], ls[3]) && isconnected(ls[3], ls[2])
true

Jusdl.Connections.release — Function

release(link::Link)

Release all the slave links of link. That is, all the slave links of link is disconnected.

Jusdl.Connections.launch — Function

launch(link::Link)

Constructs a taker task and binds it to link.

launch(link:Link)

Constructs a putter task and binds it to link.