Callback

Callbacks are used to monitor the existence of a specific events and if that specific event occurs, some other special jobs are invoked. Callbacks are intended to provide additional monitoring capability to any user-defined composite types. As such, Callbacks are generaly fields of user defined composite types objects. When a Callback is called, if the Callback is enabled and its condition function returns true, then its action function is invoked.

A Simple Example

Let's define a test object first that has a field named x of type Int and named callback of type Callback.

julia> mutable struct TestObject
       x::Int
       callback::Callback
       end

To construct an instance of TestObject, we need to construct a Callback. For that purpose, condition and action function must be defined. For this example, condition checks whether the x field is positive, and action prints a simple message saying that the x field is positive.

julia> condition(testobject) = testobject.x > 0 
condition (generic function with 1 method)

julia> action(testobject) = println("testobject.x is greater than zero") 
action (generic function with 1 method)

Now a test object can be constructed

julia> testobject = TestObject(-1, Callback(condition, action))  
TestObject(-1, Callback{typeof(condition),typeof(action)}(condition, action, true, "dac6f9eb-6daa-4622-a8fa-623f0f88780c"))

If the callback is called, no action is performed since the condition function returns false. Note the argument sent to the callback. The instance of the TestObject to which the callback is a bound.

julia> testobject.callback(testobject) 

Now mutate the test object so that condition returns true.

julia> testobject.x = 3   
3

Now, if the callback is called, since the condition returns true and the callback is enabled, the action is invoked.

julia> testobject.callback(testobject) 
testobject.x is greater than zero

Full API

Jusdl.Utilities.Callback — Type

Callback(condition, action)

Constructs a Callback from condition and action. The condition and action must be a single-argument functions. condition returns true if the condition it checks occurs, otherwise ite returns false. action is performs the specific action for which the Callback the callback is contructed. A Callback can be called by passing its single argument which is mostly bound to the Callback.

Example

julia> struct Object  # Define a dummy type.
       x::Int 
       clb::Callback 
       end 

julia> cond(obj) = obj.x > 0  # Define callback condition.
cond (generic function with 1 method)

julia> action(obj) = println("Printing the object ", obj) # Define callback action.
action (generic function with 1 method)

julia> obj = Object(1, Callback(cond, action))  # Construct an `Object` instance with `Callback`.
Object(1, Callback(condition:cond, action:action))

julia> obj.clb(obj)  # Call the callback bound `obj`.
Printing the object Object(1, Callback(condition:cond, action:action))

Jusdl.Utilities.enable! — Function

enable!(clb::Callback)

Enables clb.

Jusdl.Utilities.disable! — Function

disable!(clb::Callback)

Disables clb.

Jusdl.Utilities.isenabled — Function

isenabled(clb::Callback)

Returns true if clb is enabled. Otherwise, returns false.

Jusdl.Utilities.addcallback — Function

addcallback(obj, clb::Callback, priority::Int)

Adds clb to callback vector of obj which is assumed the have a callback list which is a vector of callback.

Example

julia> mutable struct Object 
       x::Int 
       callbacks::Vector{Callback}
       Object(x::Int) = new(x, Callback[])
       end 

julia> obj = Object(5)
Object(5, Callback[])

julia> condition(val) = val.x == 5
condition (generic function with 1 method)

julia> action(val) = @show val.x 
action (generic function with 1 method)

julia> addcallback(obj, Callback(condition, action))
Object(5, Callback[Callback(condition:condition, action:action)])

julia> obj.callbacks(obj)
val.x = 5

Jusdl.Utilities.deletecallback — Function

deletecallback(obj, idx::Int)

Deletes the one of the callbacks of obj at index idx.

julia> struct Object 
       x::Int 
       callbacks::Vector{Callback}
       end

julia> clb1 = Callback(val -> true, val -> nothing);

julia> clb2 = Callback(val -> false, val -> nothing);

julia> obj = Object(5, [clb1, clb2]);

julia> deletecallback(obj, 2);

julia> length(obj.callbacks) == 1
true