Component Type Hierarchy

Causal.AbstractComponent — Type

abstract type AbstractComponent

Abstract type of all components

Causal.AbstractSource — Type

abstract type AbstractSource <: AbstractComponent

Abstract typeof all source components

Causal.AbstractSystem — Type

abstract type AbstractSystem <: AbstractComponent

Abstract type of all system components

Causal.AbstractSink — Type

abstract type AbstractSink <: AbstractComponent

Abstract type of all sink components

Causal.AbstractStaticSystem — Type

abstract type AbstractStaticSystem <: AbstractSystem

Abstract type of all static systems

Causal.AbstractDynamicSystem — Type

abstract type AbstractDynamicSystem <: AbstractSystem

Abstract type of all dynamic system components

Causal.AbstractSubSystem — Type

abstract type AbstractSubSystem <: AbstractSystem

Abstract type of all subsystem components

Causal.AbstractMemory — Type

abstract type AbstractMemory <: AbstractStaticSystem

Abstract type of all memory components

Causal.AbstractDiscreteSystem — Type

abstract type AbstractDiscreteSystem <: AbstractDynamicSystem

Abstract type of all dynamic systems modelled by dicrete difference equations.

Causal.AbstractODESystem — Type

abstract type AbstractODESystem <: AbstractDynamicSystem

Abstract type of all dynamical systems modelled by ordinary differential equations.

Causal.AbstractRODESystem — Type

abstract type AbstractRODESystem <: AbstractDynamicSystem

Abstract type of all dynamical systems modelled by random ordinary differential equations.

Causal.AbstractDAESystem — Type

abstract type AbstractDAESystem <: AbstractDynamicSystem

Abstract type of all dynamical systems modelled by differential algebraic equations

Causal.AbstractSDESystem — Type

abstract type AbstractSDESystem <: AbstractDynamicSystem

Abstract type of all dynamical systems modelled by stochastic differential equations.

Causal.AbstractDDESystem — Type

abstract type AbstractDDESystem <: AbstractDynamicSystem

Abstract type of all dynamical systems modlled by delay dynamical systems.