Interpolation

Full API

Causal.Interpolant — Type

Interpolant(tinit, tfinal, coefinit, coeffinal)

Constructs a linnear interpolant that interpolates between the poinsts (tinit, coefinit) and (tfinal, coeffinal).

Causal.update! — Method

update!(intepolant)

Updates interpolant using the data in timebuf and databuf of interpolant.