Scope

Basic Operation of Scopes

See Basic Operation of Writers since the operation of Writer and that of Scope is very similar.

Full API

Jusdl.Components.Sinks.Scope — Type

Scope(input::Bus, buflen::Int=64, plugin=nothing, args...; kwargs...)

Constructs a Scope with input bus input. buflen is the length of the internal buffer of Scope. plugin is the additional data processing tool. args,kwargs are passed into plots(args...; kwargs...)). See (https://github.com/JuliaPlots/Plots.jl) for more information.

Warning

When initialized, the plot of Scope is closed. See open(sink::Scope) and close(sink::Scope).

Jusdl.Components.ComponentsBase.update! — Method

update!(s::Scope, x, yi)

Updates the series of the plot windows of s with x and yi.

Base.close — Method

close(sink::Scope)

Closes the plot window of the plot of sink.

Base.open — Method

open(sink::Scope)

Opens the plot window for the plots of sink.