DiscreteSystem

Construction of DiscreteSystem

DiscreteSystems evolve by the following discrete time difference equation.

\[ x_{k + 1} = f(x_k, u_k, k) \\ y_k = g(x_k, u_k, k)\]

where $x_k$ is the state, $y_k$ is the value of output and $u_k$ is the value of input at discrete time t. $f$ is the state function and $g$ is the output function of the system. See the main constructor.

Basic Construction of DiscreteSystem

When a DiscreteSystem is triggered from its trigger link, it reads current time from its trigger link, reads its input, solves its difference equation, computes its output and writes its output value to its output bus. Let us continue with an example.

We first define state function sfunc and output function ofunc of the system,

julia> using Jusdl # hide

julia> sfunc(dx, x, u, t) = (dx .= -0.5x)
sfunc (generic function with 1 method)

julia> ofunc(x, u, t) = x
ofunc (generic function with 1 method)

From sfunc, it is seen that the system does not have any input, and from ofunc the system has one output. Thus, the input and output of the system is

julia> input = nothing

julia> output = Outport(1)
1-element Outport{Outpin{Float64}}:
 Outpin(eltype:Float64, isbound:false)

We also need to specify the initial condition and time of the system

julia> x0  = [1.]
1-element Array{Float64,1}:
 1.0

julia> t = 0.
0.0

We are now ready to construct the system ds.

julia> ds = DiscreteSystem(sfunc, ofunc, x0, t, input, output)
ERROR: MethodError: no method matching DiscreteSystem(::typeof(Main.ex-discrete_system_ex.sfunc), ::typeof(Main.ex-discrete_system_ex.ofunc), ::Array{Float64,1}, ::Float64, ::Nothing, ::Outport{Outpin{Float64}})
Closest candidates are:
  DiscreteSystem(::RH, ::RO, ::ST, ::IP, ::OP, ::TR, !Matched::HS, !Matched::CB, !Matched::Symbol, !Matched::ID, !Matched::Float64, !Matched::MA, !Matched::MK, !Matched::SA, !Matched::SK, !Matched::AL, !Matched::IT) where {RH, RO, ST, IP, OP, TR, HS, CB, ID, MA, MK, SA, SK, AL, IT} at /home/sari/.julia/dev/Jusdl/src/components/systems/dynamicalsystems/discretesystems.jl:39

To drive ds, we need to launch it.

julia> iport, trg, hnd = Inport(1), Outpin(), Inpin{Bool}()
(Inport(numpins:1, eltype:Inpin{Float64}), Outpin(eltype:Float64, isbound:false), Inpin(eltype:Bool, isbound:false))

julia> connect!(ds.output, iport)
ERROR: UndefVarError: ds not defined

julia> connect!(trg, ds.trigger)
ERROR: UndefVarError: ds not defined

julia> connect!(ds.handshake, hnd)
ERROR: UndefVarError: ds not defined

julia> task = launch(ds)
ERROR: UndefVarError: ds not defined

julia> task2 = @async while true
           all(take!(iport) .=== NaN) && break
           end
Task (failed) @0x00007f959eab3820
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:97 [inlined]
_broadcast_getindex_evalf at ./broadcast.jl:631 [inlined]
_broadcast_getindex at ./broadcast.jl:604 [inlined]
getindex at ./broadcast.jl:564 [inlined]
macro expansion at ./broadcast.jl:910 [inlined]
macro expansion at ./simdloop.jl:77 [inlined]
copyto! at ./broadcast.jl:909 [inlined]
copyto! at ./broadcast.jl:864 [inlined]
copy at ./broadcast.jl:840 [inlined]
materialize at ./broadcast.jl:820 [inlined]
take!(::Inport{Inpin{Float64}}) at /home/sari/.julia/dev/Jusdl/src/connections/port.jl:178
macro expansion at ./none:2 [inlined]
(::Main.ex-discrete_system_ex.var"#1#2")() at ./task.jl:358

At this point, ds is ready to be driven. To drive ds, we can either use drive(ds, t) or put!(ds.trigger, t).

julia> put!(trg, 1.)

When the above code is executed, ds evolves until its time is ds.t is 1., During this evolution, ds reads time t from its trigger link, reads its input (in this example, ds has no input, so it does nothing when reading its input), solves its difference equation, computes its output and writes its output value to its output. To signal that the evolution is succeeded, ds writes true its handshake link which needs to be taken to further drive ds.

julia> hnd.link  # `handshake` link is readable
ERROR: UndefRefError: access to undefined reference

julia> take!(hnd)
ERROR: UndefRefError: access to undefined reference

We continue to drive ds,

julia> for i in 2. : 10.
           put!(trg, i)
           take!(hnd)
       end
ERROR: UndefRefError: access to undefined reference

Note that all the output values of ds is written to its output bus.

julia> iport[1].link.buffer
ERROR: UndefRefError: access to undefined reference

When we launched ds, we constructed a task which is still running.

julia> task
ERROR: UndefVarError: task not defined

julia> task2
Task (failed) @0x00007f959eab3820
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:97 [inlined]
_broadcast_getindex_evalf at ./broadcast.jl:631 [inlined]
_broadcast_getindex at ./broadcast.jl:604 [inlined]
getindex at ./broadcast.jl:564 [inlined]
macro expansion at ./broadcast.jl:910 [inlined]
macro expansion at ./simdloop.jl:77 [inlined]
copyto! at ./broadcast.jl:909 [inlined]
copyto! at ./broadcast.jl:864 [inlined]
copy at ./broadcast.jl:840 [inlined]
materialize at ./broadcast.jl:820 [inlined]
take!(::Inport{Inpin{Float64}}) at /home/sari/.julia/dev/Jusdl/src/connections/port.jl:178
macro expansion at ./none:2 [inlined]
(::Main.ex-discrete_system_ex.var"#1#2")() at ./task.jl:358

As long nothing goes wrong, i.e. no exception is thrown, during the evolution of ds, it is possible to drive ds. To safely terminate the task, we need to terminate the ds.

julia> put!(trg, NaN)

julia> put!(ds.output, [NaN])
ERROR: UndefVarError: ds not defined

We can confirm that the task is not running and its state is done.

julia> task
ERROR: UndefVarError: task not defined

julia> task2
Task (failed) @0x00007f959eab3820
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:97 [inlined]
_broadcast_getindex_evalf at ./broadcast.jl:631 [inlined]
_broadcast_getindex at ./broadcast.jl:604 [inlined]
getindex at ./broadcast.jl:564 [inlined]
macro expansion at ./broadcast.jl:910 [inlined]
macro expansion at ./simdloop.jl:77 [inlined]
copyto! at ./broadcast.jl:909 [inlined]
copyto! at ./broadcast.jl:864 [inlined]
copy at ./broadcast.jl:840 [inlined]
materialize at ./broadcast.jl:820 [inlined]
take!(::Inport{Inpin{Float64}}) at /home/sari/.julia/dev/Jusdl/src/connections/port.jl:178
macro expansion at ./none:2 [inlined]
(::Main.ex-discrete_system_ex.var"#1#2")() at ./task.jl:358

Since the task is not running any more, ds cannot be drivable any more. However to drive ds again, we need launch ds again.

Full API

Jusdl.@def_discrete_system — Macro

@def_discrete_system

Used to define discrete time system

Jusdl.DiscreteSystem — Type

DiscreteSystem(; righthandside, readout, state, input, output)

Constructs a generic discrete system

Jusdl.DiscreteLinearSystem — Type

DiscreteLinearSystem(input, output, modelargs=(), solverargs=(); 
    A=fill(-1, 1, 1), B=fill(0, 1, 1), C=fill(1, 1, 1), D=fill(0, 1, 1), state=rand(size(A,1)), t=0., 
    alg=ODEAlg, modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a DiscreteLinearSystem with input and output. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

The DiscreteLinearSystem is represented by the following state and output equations.

\[\begin{array}{l} \dot{x} = A x + B u \\[0.25cm] y = C x + D u \end{array}\]

where $x$ is state. solver is used to solve the above differential equation.

Jusdl.HenonSystem — Type

Henon()

Constructs a Henon system evolving with the dynamics

\[\begin{array}{l} \dot{x}_1 = 1 - \alpha (x_1^2) + x_2 \\[0.25cm] \dot{x}_2 = \beta x_1 \end{array}\]

Jusdl.LoziSystem — Type

LoziSystem()

Constructs a Lozi system evolving with the dynamics

\[\begin{array}{l} \dot{x}_1 = 1 - \alpha |x_1| + x_2 \\[0.25cm] \dot{x}_2 = \beta x_1 \end{array}\]

Jusdl.BogdanovSystem — Type

BogdanovSystem()

Constructs a Bogdanov system with equations

\[\begin{array}{l} \dot{x}_1 = x_1 + \dot{x}_2 \\[0.25cm] \dot{x}_2 = x_2 + \epsilon + x_2 + k x_1 (x_1 - 1) + \mu x_1 x_2 \end{array}\]

Jusdl.GingerbreadmanSystem — Type

GingerbreadmanSystem()

Constructs a GingerbreadmanSystem with the dynamics

\[\begin{array}{l} \dot{x}_1 = 1 - x_2 + |x_1|\\[0.25cm] \dot{x}_2 = x_1 \end{array}\]

Jusdl.LogisticSystem — Type

LogisticSystem()

Constructs a LogisticSystem with the dynamics

\[\begin{array}{l} \dot{x} = r x (1 - x) \end{array}\]