ODESystem

Basic Operation of ODESystem

When an ODESystem is triggered, it reads its current time from its trigger link, reads its input, solves its differential equation and computes its output. Let us observe the basic operation of ODESystems with a simple example.

We first construct an ODESystem. Since an ODESystem is represented by its state equation and output equation, we need to define those equations.

julia> using Jusdl # hide

julia> sfunc(dx,x,u,t) = (dx .= -0.5x)
sfunc (generic function with 1 method)

julia> ofunc(x, u, t) = x
ofunc (generic function with 1 method)

Let us construct the system

julia> ds = ODESystem(sfunc, ofunc, [1.], 0., Inport(1), Outport(1))
ODESystem(state:[1.0], t:0.0, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

Note that ds is a single input single output ODESystem with an initial state of [1.] and initial time 0.. To drive, i.e. trigger ds, we need to launch it.

julia> task = launch(ds)
Task (failed) @0x00007f59c3c2c760
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:74 [inlined]
readtime at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:12 [inlined]
takestep at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:129 [inlined]
macro expansion at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:172 [inlined]
(::Jusdl.var"#33#34"{ODESystem{typeof(Main.ex-ode_ex.sfunc),typeof(Main.ex-ode_ex.ofunc),Array{Float64,1},Float64,OrdinaryDiffEq.ODEIntegrator{OrdinaryDiffEq.Tsit5,true,Array{Float64,1},Nothing,Float64,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},Float64,Float64,Float64,Array{Array{Float64,1},1},DiffEqBase.ODESolution{Float64,2,Array{Array{Float64,1},1},Nothing,Nothing,Array{Float64,1},Array{Array{Array{Float64,1},1},1},DiffEqBase.ODEProblem{Array{Float64,1},Tuple{Float64,Float64},true,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}},DiffEqBase.StandardODEProblem},OrdinaryDiffEq.Tsit5,OrdinaryDiffEq.InterpolationData{DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Array{Float64,1},1},Array{Float64,1},Array{Array{Array{Float64,1},1},1},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}}},DiffEqBase.DEStats},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}},OrdinaryDiffEq.DEOptions{Float64,Float64,Float64,Float64,typeof(DiffEqBase.ODE_DEFAULT_NORM),typeof(LinearAlgebra.opnorm),DiffEqBase.CallbackSet{Tuple{},Tuple{}},typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN),typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE),typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK),DataStructures.BinaryHeap{Float64,DataStructures.LessThan},DataStructures.BinaryHeap{Float64,DataStructures.LessThan},Nothing,Nothing,Int64,Array{Float64,1},Array{Float64,1},Array{Float64,1}},Array{Float64,1},Float64,Nothing,OrdinaryDiffEq.DefaultInit},Inport{Inpin{Float64}},Outport{Outpin{Float64}},Inpin{Float64},Outpin{Bool},Nothing}})() at ./task.jl:358

When launced, ds is ready to driven. ds is driven from its trigger link. Note that the trigger link of ds is writable.

julia> ds.trigger
Inpin(eltype:Float64, isbound:false)

Let us drive ds to the time of t of 1 second.

julia> drive(ds, 1.)
ERROR: MethodError: no method matching put!(::Inpin{Float64}, ::Float64)
Closest candidates are:
  put!(!Matched::Channel{T}, ::Any) where T at channels.jl:312
  put!(!Matched::Distributed.Future, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:557
  put!(!Matched::Distributed.RemoteValue, ::Any...) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:572
  ...

When driven, ds reads current time of t from its trigger link, reads its input value from its input, solves its differential equation and computes its output values and writes its output. So, for the step to be continued, an input values must be written. Note that the input of ds is writable,

julia> ds.input
1-element Inport{Inpin{Float64}}:
 Inpin(eltype:Float64, isbound:false)

Let us write some value.

julia> put!(ds.input, [5.])
ERROR: MethodError: no method matching put!(::Inport{Inpin{Float64}}, ::Array{Float64,1})
Closest candidates are:
  put!(!Matched::Channel{T}, ::Any) where T at channels.jl:312
  put!(!Matched::Distributed.Future, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:557
  put!(!Matched::Distributed.RemoteValue, ::Any...) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:572
  ...

At this point, ds completed its step and put true to its handshake link to signal that its step is succeeded.

julia> ds.handshake
Outpin(eltype:Bool, isbound:false)

To complete the step and be ready for another step, we need to approve the step by reading its handshake.

julia> take!(ds.handshake)
ERROR: MethodError: no method matching take!(::Outpin{Bool})
Closest candidates are:
  take!(!Matched::Distributed.WorkerPool) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/workerpool.jl:137
  take!(!Matched::IOStream) at iostream.jl:383
  take!(!Matched::Base.GenericIOBuffer{Array{UInt8,1}}) at iobuffer.jl:381
  ...

At this point, ds can be driven further.

julia> for t in 2. : 10.
           put!(ds.trigger, t)
           put!(ds.input, [t * 10])
           take!(ds.handshake)
       end
ERROR: MethodError: no method matching put!(::Inpin{Float64}, ::Float64)
Closest candidates are:
  put!(!Matched::Channel{T}, ::Any) where T at channels.jl:312
  put!(!Matched::Distributed.Future, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:557
  put!(!Matched::Distributed.RemoteValue, ::Any...) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:572
  ...

Note that all the output value of ds is written to its outputbus,

julia> ds.output[1].buffer.data
ERROR: type Outpin has no field buffer

When we launched ds, we constructed a task and the task is still running.

julia> task
Task (failed) @0x00007f59c3c2c760
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:74 [inlined]
readtime at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:12 [inlined]
takestep at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:129 [inlined]
macro expansion at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:172 [inlined]
(::Jusdl.var"#33#34"{ODESystem{typeof(Main.ex-ode_ex.sfunc),typeof(Main.ex-ode_ex.ofunc),Array{Float64,1},Float64,OrdinaryDiffEq.ODEIntegrator{OrdinaryDiffEq.Tsit5,true,Array{Float64,1},Nothing,Float64,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},Float64,Float64,Float64,Array{Array{Float64,1},1},DiffEqBase.ODESolution{Float64,2,Array{Array{Float64,1},1},Nothing,Nothing,Array{Float64,1},Array{Array{Array{Float64,1},1},1},DiffEqBase.ODEProblem{Array{Float64,1},Tuple{Float64,Float64},true,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}},DiffEqBase.StandardODEProblem},OrdinaryDiffEq.Tsit5,OrdinaryDiffEq.InterpolationData{DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Array{Float64,1},1},Array{Float64,1},Array{Array{Array{Float64,1},1},1},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}}},DiffEqBase.DEStats},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}},OrdinaryDiffEq.DEOptions{Float64,Float64,Float64,Float64,typeof(DiffEqBase.ODE_DEFAULT_NORM),typeof(LinearAlgebra.opnorm),DiffEqBase.CallbackSet{Tuple{},Tuple{}},typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN),typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE),typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK),DataStructures.BinaryHeap{Float64,DataStructures.LessThan},DataStructures.BinaryHeap{Float64,DataStructures.LessThan},Nothing,Nothing,Int64,Array{Float64,1},Array{Float64,1},Array{Float64,1}},Array{Float64,1},Float64,Nothing,OrdinaryDiffEq.DefaultInit},Inport{Inpin{Float64}},Outport{Outpin{Float64}},Inpin{Float64},Outpin{Bool},Nothing}})() at ./task.jl:358

To terminate the task safely, we need to terminate ds safely.

julia> terminate(ds)
ERROR: MethodError: no method matching close(::Outport{Outpin{Float64}})
Closest candidates are:
  close(!Matched::Base.DevNull) at coreio.jl:18
  close(!Matched::TextDisplay) at multimedia.jl:252
  close(!Matched::Base.Libc.FILE) at libc.jl:93
  ...

Now, the state of the task is done.

julia> task
Task (failed) @0x00007f59c3c2c760
UndefRefError: access to undefined reference
getproperty at ./Base.jl:33 [inlined]
take! at /home/sari/.julia/dev/Jusdl/src/connections/pin.jl:74 [inlined]
readtime at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:12 [inlined]
takestep at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:129 [inlined]
macro expansion at /home/sari/.julia/dev/Jusdl/src/components/componentsbase/takestep.jl:172 [inlined]
(::Jusdl.var"#33#34"{ODESystem{typeof(Main.ex-ode_ex.sfunc),typeof(Main.ex-ode_ex.ofunc),Array{Float64,1},Float64,OrdinaryDiffEq.ODEIntegrator{OrdinaryDiffEq.Tsit5,true,Array{Float64,1},Nothing,Float64,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},Float64,Float64,Float64,Array{Array{Float64,1},1},DiffEqBase.ODESolution{Float64,2,Array{Array{Float64,1},1},Nothing,Nothing,Array{Float64,1},Array{Array{Array{Float64,1},1},1},DiffEqBase.ODEProblem{Array{Float64,1},Tuple{Float64,Float64},true,Interpolant{Buffer{Cyclic,Float64,1},Buffer{Cyclic,Float64,1},Array{Interpolations.Extrapolation{Float64,1,Interpolations.ScaledInterpolation{Float64,1,Interpolations.BSplineInterpolation{Float64,1,OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{Base.OneTo{Int64}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}},Interpolations.BSpline{Interpolations.Cubic{Interpolations.Line{Interpolations.OnGrid}}},Interpolations.Line{Nothing}},1}},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}},DiffEqBase.StandardODEProblem},OrdinaryDiffEq.Tsit5,OrdinaryDiffEq.InterpolationData{DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},Array{Array{Float64,1},1},Array{Float64,1},Array{Array{Array{Float64,1},1},1},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}}},DiffEqBase.DEStats},DiffEqBase.ODEFunction{true,typeof(Main.ex-ode_ex.sfunc),LinearAlgebra.UniformScaling{Bool},Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing,Nothing},OrdinaryDiffEq.Tsit5Cache{Array{Float64,1},Array{Float64,1},Array{Float64,1},OrdinaryDiffEq.Tsit5ConstantCache{Float64,Float64}},OrdinaryDiffEq.DEOptions{Float64,Float64,Float64,Float64,typeof(DiffEqBase.ODE_DEFAULT_NORM),typeof(LinearAlgebra.opnorm),DiffEqBase.CallbackSet{Tuple{},Tuple{}},typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN),typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE),typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK),DataStructures.BinaryHeap{Float64,DataStructures.LessThan},DataStructures.BinaryHeap{Float64,DataStructures.LessThan},Nothing,Nothing,Int64,Array{Float64,1},Array{Float64,1},Array{Float64,1}},Array{Float64,1},Float64,Nothing,OrdinaryDiffEq.DefaultInit},Inport{Inpin{Float64}},Outport{Outpin{Float64}},Inpin{Float64},Outpin{Bool},Nothing}})() at ./task.jl:358

So, it is not possible to drive ds.

Mutation in State Function in ODESystem

Consider a system with the following ODE

\[\begin{array}{l} \dot{x} = f(x, u, t) \\ y = g(x, u, t) \\ \end{array}\]

where $x \in R^d, y \in R^m, u \in R^p$. To construct and ODESystem, The signature of the state function statefunc must be of the form

function statefunc(dx, x, u, t)
    dx .= ... # Update dx
end

Note that statefunc does not construct dx but updates dx and does not return anything. This is for performance reasons. On the contrary, the signature of the output function outputfunc must be of the form,

function outputfunc(x, u, t)
    y = ... # Compute y
    return y
end

Note the output value y is computed and returned from outputfunc. y is not updated but generated in the outputfunc.

Full API

Jusdl.ODESystem — Type

ODESystem(input, output, statefunc, outputfunc, state, t, modelargs=(), solverargs=(); 
    alg=ODEAlg, modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs an ODESystem with input and output. statefunc is the state function and outputfunc is the output function. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

ODESystem is represented by the equations.

\[ \begin{array}{l} \dot{x} = f(x, u, t) \\[0.25cm] y = g(x, u, t) \end{array}\]

where $t$ is the time t, $x$ is state, $u$ is the value of input, $y$ is the value of output. $f$ is statefunc and $g$ is outputfunc. solver is used to solve the above differential equation.

The signature of statefunc must be of the form,

function statefunc(dx, x, u, t)
    dx .= ... # Update dx 
end

and the signature of outputfunc must be of the form,

function outputfunc(x, u, t)
    y = ... # Compute y
    return y
end

Example

julia> sfuncode(dx,x,u,t) = (dx[1] = 0.5x[1] + u[1](t));

julia> ofuncode(x, u, t) = x;

julia> ds = ODESystem(sfuncode, ofuncode, [1.], 0., Inport(), Outport())
ODESystem(state:[1.0], t:0.0, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

julia> ds = ODESystem(sfuncode, ofuncode, [1.], 0., Inport(), Outport(), solverkwargs=(dt=0.1, reltol=1e-6))
ODESystem(state:[1.0], t:0.0, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

Info

See DifferentialEquations for more information about modelargs, modelkwargs, solverargs, solverkwargs and alg.

Jusdl.LinearSystem — Type

LinearSystem(input, output, modelargs=(), solverargs=(); 
    A=fill(-1, 1, 1), B=fill(0, 1, 1), C=fill(1, 1, 1), D=fill(0, 1, 1), state=rand(size(A,1)), t=0., 
    alg=ODEAlg, modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a LinearSystem with input and output. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

The LinearSystem is represented by the following state and output equations.

\[\begin{array}{l} \dot{x} = A x + B u \\[0.25cm] y = C x + D u \end{array}\]

where $x$ is state. solver is used to solve the above differential equation.

Jusdl.LorenzSystem — Type

LorenzSystem(input, output, modelargs=(), solverargs=(); 
    sigma=10, beta=8/3, rho=28, gamma=1, outputfunc=allstates, state=rand(3), t=0.,
    alg=ODEAlg, cplmat=diagm([1., 1., 1.]), modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a LorenzSystem with input and output. sigma, beta, rho and gamma is the system parameters. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

If input is nothing, the state equation of LorenzSystem is

\[\begin{array}{l} \dot{x}_1 = \gamma (\sigma (x_2 - x_1)) \\[0.25cm] \dot{x}_2 = \gamma (x_1 (\rho - x_3) - x_2) \\[0.25cm] \dot{x}_3 = \gamma (x_1 x_2 - \beta x_3) \end{array}\]

where $x$ is state. solver is used to solve the above differential equation. If input is not nothing, then the state eqaution is

\[\begin{array}{l} \dot{x}_1 = \gamma (\sigma (x_2 - x_1)) + \sum_{j = 1}^3 \alpha_{1j} u_j \\[0.25cm] \dot{x}_2 = \gamma (x_1 (\rho - x_3) - x_2) + \sum_{j = 1}^3 \alpha_{2j} u_j \\[0.25cm] \dot{x}_3 = \gamma (x_1 x_2 - \beta x_3) + \sum_{j = 1}^3 \alpha_{3j} u_j \end{array}\]

where $A = [\alpha_{ij}]$ is cplmat and $u = [u_{j}]$ is the value of the input. The output function is

\[ y = g(x, u, t)\]

where $t$ is time t, $y$ is the value of the output and $g$ is outputfunc.

Jusdl.ChenSystem — Type

ChenSystem(input, output, modelargs=(), solverargs=(); 
    a=35, b=3, c=28, gamma=1, outputfunc=allstates, state=rand(3), t=0.,
    alg=ODEAlg, cplmat=diagm([1., 1., 1.]), modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a ChenSystem with input and output. a, b, c and gamma is the system parameters. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

If input is nothing, the state equation of ChenSystem is

\[\begin{array}{l} \dot{x}_1 = \gamma (a (x_2 - x_1)) \\[0.25cm] \dot{x}_2 = \gamma ((c - a) x_1 + c x_2 + x_1 x_3) \\[0.25cm] \dot{x}_3 = \gamma (x_1 x_2 - b x_3) \end{array}\]

where $x$ is state. solver is used to solve the above differential equation. If input is not nothing, then the state eqaution is

\[\begin{array}{l} \dot{x}_1 = \gamma (a (x_2 - x_1)) + \sum_{j = 1}^3 \alpha_{1j} u_j \\[0.25cm] \dot{x}_2 = \gamma ((c - a) x_1 + c x_2 + x_1 x_3) + \sum_{j = 1}^3 \alpha_{2j} u_j \\[0.25cm] \dot{x}_3 = \gamma (x_1 x_2 - b x_3) + \sum_{j = 1}^3 \alpha_{3j} u_j \end{array}\]

where $A = [\alpha_{ij}]$ is cplmat and $u = [u_{j}]$ is the value of the input. The output function is

\[ y = g(x, u, t)\]

where $t$ is time t, $y$ is the value of the output and $g$ is outputfunc.

Jusdl.ChuaSystem — Type

ChuaSystem(input, output, modelargs=(), solverargs=(); 
    diode=PiecewiseLinearDiode(), alpha=15.6, beta=28., gamma=1., outputfunc=allstates, state=rand(3), t=0., 
    alg=ODEAlg, cplmat=diagm([1., 1., 1.]), modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a ChuaSystem with input and output. diode, alpha, beta and gamma is the system parameters. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

If input is nothing, the state equation of ChuaSystem is

\[\begin{array}{l} \dot{x}_1 = \gamma (\alpha (x_2 - x_1 - h(x_1))) \\[0.25cm] \dot{x}_2 = \gamma (x_1 - x_2 + x_3 ) \\[0.25cm] \dot{x}_3 = \gamma (-\beta x_2) \end{array}\]

where $x$ is state. solver is used to solve the above differential equation. If input is not nothing, then the state eqaution is

\[\begin{array}{l} \dot{x}_1 = \gamma (\alpha (x_2 - x_1 - h(x_1))) + \sum_{j = 1}^3 \theta_{1j} u_j \\[0.25cm] \dot{x}_2 = \gamma (x_1 - x_2 + x_3 ) + \sum_{j = 1}^3 \theta_{2j} u_j \\[0.25cm] \dot{x}_3 = \gamma (-\beta x_2) + \sum_{j = 1}^3 \theta_{3j} u_j \end{array}\]

where $\Theta = [\theta_{ij}]$ is cplmat and $u = [u_{j}]$ is the value of the input. The output function is

\[ y = g(x, u, t)\]

where $t$ is time t, $y$ is the value of the output and $g$ is outputfunc.

Jusdl.RosslerSystem — Type

RosslerSystem(input, output, modelargs=(), solverargs=(); 
    a=0.38, b=0.3, c=4.82, gamma=1., outputfunc=allstates, state=rand(3), t=0., 
    alg=ODEAlg, cplmat=diagm([1., 1., 1.]), modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a RosllerSystem with input and output. a, b, c and gamma is the system parameters. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

If input is nothing, the state equation of RosslerSystem is

\[\begin{array}{l} \dot{x}_1 = \gamma (-x_2 - x_3) \\[0.25cm] \dot{x}_2 = \gamma (x_1 + a x_2) \\[0.25cm] \dot{x}_3 = \gamma (b + x_3 (x_1 - c)) \end{array}\]

where $x$ is state. solver is used to solve the above differential equation. If input is not nothing, then the state eqaution is

\[\begin{array}{l} \dot{x}_1 = \gamma (-x_2 - x_3) + \sum_{j = 1}^3 \theta_{1j} u_j \\[0.25cm] \dot{x}_2 = \gamma (x_1 + a x_2 ) + \sum_{j = 1}^3 \theta_{2j} u_j \\[0.25cm] \dot{x}_3 = \gamma (b + x_3 (x_1 - c)) + \sum_{j = 1}^3 \theta_{3j} u_j \end{array}\]

where $\Theta = [\theta_{ij}]$ is cplmat and $u = [u_{j}]$ is the value of the input. The output function is

\[ y = g(x, u, t)\]

where $t$ is time t, $y$ is the value of the output and $g$ is outputfunc.

Jusdl.VanderpolSystem — Type

VanderpolSystem(input, output, modelargs=(), solverargs=(); 
    mu=5., gamma=1., outputfunc=allstates, state=rand(2), t=0., 
    alg=ODEAlg, cplmat=diagm([1., 1]), modelkwargs=NamedTuple(), solverkwargs=NamedTuple())

Constructs a VanderpolSystem with input and output. mu and gamma is the system parameters. state is the initial state and t is the time. modelargs and modelkwargs are passed into ODEProblem and solverargs and solverkwargs are passed into solve method of DifferentialEquations. alg is the algorithm to solve the differential equation of the system.

If input is nothing, the state equation of VanderpolSystem is

\[\begin{array}{l} \dot{x}_1 = \gamma (x_2) \\[0.25cm] \dot{x}_2 = \gamma (\mu (x_1^2 - 1) x_2 - x_1 ) \end{array}\]

where $x$ is state. solver is used to solve the above differential equation. If input is not nothing, then the state eqaution is

\[\begin{array}{l} \dot{x}_1 = \gamma (x_2) + \sum_{j = 1}^3 \theta_{1j} u_j \\[0.25cm] \dot{x}_2 = \gamma (\mu (x_1^2 - 1) x_2 - x_1) + \sum_{j = 1}^3 \theta_{2j} u_j \end{array}\]

where $\Theta = [\theta_{ij}]$ is cplmat and $u = [u_{j}]$ is the value of the input. The output function is

\[ y = g(x, u, t)\]

where $t$ is time t, $y$ is the value of the output and $g$ is outputfunc.