Subsystem

Construction of SubSystems

A SubSystem consists of connected components. Thus, to construct a SubSystem, we first construct components, connect them and specify the input and output of SubSystem. See the basic constructor.

Jusdl.SubSystem — Type

SubSystem(components, input, output; callbacks=nothing, name=Symbol())

Constructs a SubSystem consisting of components. input and output determines the input and output of SubSystem. input and output may be nothing, a vector of pins or a port.

Example

julia> adder = Adder((+,-));

julia> gain = Gain(); 

julia> gen = ConstantGenerator();

julia> connect(gen.output, adder.input);

julia> connect(adder.output, gain.input);

julia> ss = SubSystem([gen, adder, gain], adder.input[1], gain.output)
SubSystem(input:Inpin(eltype:Float64, isbound:true), output:Outport(numpins:1, eltype:Outpin{Float64}), components:AbstractComponent[ConstantGenerator(amp:1.0), Adder(signs:(+, -), input:Inport(numpins:2, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64})), Gain(gain:1.0, input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))])

Basic Operation of SubSystems

The operation of a SubSystem is very similar to that of a StaticSystem. The only difference is that when a SubSystem is triggered through its trigger link, it distributes the trigger to the trigger links of its components. Then, each of the components of the SubSystem takes steps individually.

Let us construct a subsystem consisting of a generator and an adder.

julia> using Jusdl # hide

julia> gen = ConstantGenerator()
ConstantGenerator(amp:1.0)

julia> adder = Adder((+,+))
Adder(signs:(+, +), input:Inport(numpins:2, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))

Connect the generator and adder.

julia> connect(gen.output, adder.input[1])
1-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, isreadable:false, iswritable:false)

We are ready to construct a SubSystem.

julia> sub = SubSystem([gen, adder], [adder.input[2]], adder.output)
SubSystem(input:Inport(numpins:1, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}), components:AbstractComponent[ConstantGenerator(amp:1.0), Adder(signs:(+, +), input:Inport(numpins:2, eltype:Inpin{Float64}), output:Outport(numpins:1, eltype:Outpin{Float64}))])

To trigger the sub, we need to launch it,

julia> t = launch(sub)
3-element Array{Task,1}:
 Task (runnable) @0x00007f6ab421b340
 Task (runnable) @0x00007f6ab42509d0
 Task (failed) @0x00007f6ab4219d50

sub is ready to be triggered,

julia> drive(sub, 1.)
ERROR: MethodError: no method matching put!(::Inpin{Float64}, ::Float64)
Closest candidates are:
  put!(!Matched::Channel{T}, ::Any) where T at channels.jl:312
  put!(!Matched::Distributed.Future, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:557
  put!(!Matched::Distributed.RemoteValue, ::Any...) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:572
  ...

Put some data to the input of sub.

julia> put!(sub.input, [1.])
ERROR: MethodError: no method matching put!(::Inport{Inpin{Float64}}, ::Array{Float64,1})
Closest candidates are:
  put!(!Matched::Channel{T}, ::Any) where T at channels.jl:312
  put!(!Matched::Distributed.Future, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:557
  put!(!Matched::Distributed.RemoteValue, ::Any...) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/remotecall.jl:572
  ...

The step needs to be approved.

julia> approve(sub)
ERROR: MethodError: no method matching take!(::Outpin{Bool})
Closest candidates are:
  take!(!Matched::Distributed.WorkerPool) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.4/Distributed/src/workerpool.jl:137
  take!(!Matched::IOStream) at iostream.jl:383
  take!(!Matched::Base.GenericIOBuffer{Array{UInt8,1}}) at iobuffer.jl:381
  ...

Now print the data written to the outputs of the components of sub.

julia> sub.components[1].output[1].buffer.data[1]
ERROR: type Outpin has no field buffer

julia> sub.components[2].output[1].buffer.data[1]
ERROR: type Outpin has no field buffer

Note that when sub is triggered, sub transfer the trigger to all its internal components.