Ports

A Port is actually is a bunch of pins. Reading from and writing into data is performed as in the case of Links.

Construction of Ports

A Port is constructed by specifying its element type T, the number of pins npins and the buffer length of its pins.

Missing docstring.

Missing docstring for Outport. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Inport. Check Documenter's build log for details.

Connection and Disconnection of Ports

The Portsses can be connected to and disconnected from each other. When connected any data written to the master port is also written all slave ports. See the following example.

Let us connect two ports and connect them together.

julia> op = Outport(2)  # Port with `2` pins with buffer length of `5`.
2-element Outport{Outpin{Float64}}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

julia> ip = Inport(2)  # Port with `2` pins with buffer length of `5`.
2-element Inport{Inpin{Float64}}:
 Inpin(eltype:Float64, isbound:false)
 Inpin(eltype:Float64, isbound:false)

julia> ls = connect(op, ip)
2-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, isreadable:false, iswritable:false)

Here, op is the master port and ip is the slave port. That is, data written to op is also written into ip.

julia> t1 = @async while true
           val = take!(ip)
           all(val .=== NaN) && break
           println("Took " * string(val))
       end
Task (runnable) @0x00007f59c3e26980

julia> put!(op, [5., 10.]);
Took [5.0, 10.0]

julia> [outbuf(pin.link.buffer) for pin in ip]
2-element Array{Array{Float64,1},1}:
 [5.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
 [10.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]

Note that the data [5, 10] written to op is also written ip since they are connected.

The Portses connected to each other can be disconnected. When disconnected, the data written to master is not written to slaves

julia> disconnect(op, ip)

julia> isconnected(op, ip)
false

Data Flow through Ports

Data flow through the Portses is very similar to the case in Links. See Data Flow through Links for information about data flow through Links. Runnable tasks must be bound to the pins of the ports for data flow through the Port. Again, put! and take! functions are used to write data from a Port and read from data from a Port.

Base.put! — Method

put!(outport::Outport, vals)

Puts vals to outport. Each item in vals is putted to the links of the outport.

Warning

The outport must be writable to be read. That is, there must be a runnable tasks bound to links of the outport that reads data from outport.

Example

julia> op, ip = Outport(), Inport() 
(Outport(numpins:1, eltype:Outpin{Float64}), Inport(numpins:1, eltype:Inpin{Float64}))

julia> ls = connect(op, ip)
1-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, isreadable:false, iswritable:false)

julia> t = @async while true 
       val = take!(ip)
       all(val .=== NaN) && break 
       println("Took " * string(val))
       end;

julia> put!(op, [1.])
Took [1.0]
1-element Array{Float64,1}:
 1.0

julia> put!(op, [NaN])
1-element Array{Float64,1}:
 NaN

Base.take! — Method

take!(inport::Inport)

Takes an element from inport. Each link of the inport is a read and a vector containing the results is returned.

Warning

The inport must be readable to be read. That is, there must be a runnable tasks bound to links of the inport that writes data to inport.

Example

julia> op, ip = Outport(), Inport()
(Outport(numpins:1, eltype:Outpin{Float64}), Inport(numpins:1, eltype:Inpin{Float64}))

julia> ls = connect(op, ip)
1-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, isreadable:false, iswritable:false)

julia> t = @async for val in 1 : 5 
       put!(op, [val])
       end;

julia> take!(ip)
1-element Array{Float64,1}:
 1.0

julia> take!(ip)
1-element Array{Float64,1}:
 2.0

Any data written to a Port is recorded into the buffers of its pins.

@repl writing_to_busses
using Jusdl # hide
op, ip = Outport(2), Inport(2);
t = @async while true
    val = take!(ip)
    all(val .=== NaN) && break
    println("Took " * string(val))
end
put!(op, 1.);
ip[1].link.buffer

Indexing and Iteration of Ports

Ports can be indexed similarly to the arrays in Julia. When indexed, the corresponding pin of the port is returned.

julia> using Jusdl # hide

julia> op = Outport(3)
3-element Outport{Outpin{Float64}}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

julia> op[1]
Outpin(eltype:Float64, isbound:false)

julia> op[end]
Outpin(eltype:Float64, isbound:false)

julia> op[:]
3-element Array{Outpin{Float64},1}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

julia> op[1] = Outpin()
Outpin(eltype:Float64, isbound:false)

julia> op[1:2] = [Outpin(), Outpin()]
2-element Array{Outpin{Float64},1}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

The iteration of Ports in a loop is also possible. When iterated, the pins of the Port is returned.

julia> using Jusdl # hide

julia> port = Inport(3)
3-element Inport{Inpin{Float64}}:
 Inpin(eltype:Float64, isbound:false)
 Inpin(eltype:Float64, isbound:false)
 Inpin(eltype:Float64, isbound:false)

julia> for pin in port
           @show pin
       end
pin = Inpin(eltype:Float64, isbound:false)
pin = Inpin(eltype:Float64, isbound:false)
pin = Inpin(eltype:Float64, isbound:false)

Full API

Jusdl.datatype — Method

datatype(port::AbstractPort)

Returns the data type of port.

Base.size — Method

size(port::AbstractPort)

Retruns size of port.

Base.getindex — Method

getindex(port::AbstractPort, idx::Vararg{Int, N}) where N

Returns elements from port at index idx. Same as port[idx].

Example

julia> op = Outport(3)
3-element Outport{Outpin{Float64}}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

julia> op[1]
Outpin(eltype:Float64, isbound:false)

julia> op[end]
Outpin(eltype:Float64, isbound:false)

julia> op[:]
3-element Array{Outpin{Float64},1}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

Base.setindex! — Method

setindex!(port::AbstractPort, item, idx::Vararg{Int, N}) where N

Sets item to port at index idx. Same as port[idx] = item.

Example

julia> op = Outport(3)
3-element Outport{Outpin{Float64}}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

julia> op[1] = Outpin()
Outpin(eltype:Float64, isbound:false)

julia> op[end] = Outpin()
Outpin(eltype:Float64, isbound:false)

julia> op[1:2] = [Outpin(), Outpin()]
2-element Array{Outpin{Float64},1}:
 Outpin(eltype:Float64, isbound:false)
 Outpin(eltype:Float64, isbound:false)

Missing docstring.

Missing docstring for take!(inport::Inport). Check Documenter's build log for details.

Missing docstring.

Missing docstring for put!(outport::Outport, vals). Check Documenter's build log for details.

Missing docstring.

Missing docstring for similar(outport::Outport{P}, numpins::Int=length(outport)) where {P<:Outpin{T}} where {T}. Check Documenter's build log for details.