Busses

A Bus is actually is a bunch of links. Reading from and writing into data is performed as in the case of Links.

Construction of Bus

A Bus is constructed by specifying its element type T, number of links nlinks and the buffer length of its links.

Jusdl.Connections.Bus — Type

Bus{T}([nlinks::Int=1, [ln::Int=64]]) where T

Constructs a Bus consisting of nlinks links. ln is the buffer length and T is element type of the links.

Connection and Disconnection of Busses

The Busses can be connected and disconnected to each other. When connected any data written to the master bus is also written all slave busses. See the following example.

Let us connect two busses and connect them together.

julia> b1 = Bus(2, 5)  # Bus with `2` links with buffer length of `5`.
2-element Bus{Link{Float64}}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b2 = Bus(2, 5)  # Bus with `2` links with buffer length of `5`.
2-element Bus{Link{Float64}}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> connect(b1, b2)

Here, b1 is the master bus and b2 is the slave bus. That is, data written to b1 is also written into b2.

julia> t1 = launch(b1);

julia> t2 = launch(b2);

julia> put!(b1, [5., 10.]);
Exception handling log message: Exception handling log message: Exception handling log message: BaseException handling log message: 
.IOError
julia> [b1[i].buffer.data for i = 1 : 2]
2-element Array{Array{Float64,1},1}:
 [5.0, 0.0, 0.0, 0.0, 0.0]
 [10.0, 0.0, 0.0, 0.0, 0.0]

julia> [b2[i].buffer.data for i = 1 : 2]
2-element Array{Array{Float64,1},1}:
 [5.0, 0.0, 0.0, 0.0, 0.0]
 [10.0, 0.0, 0.0, 0.0, 0.0]

Note that the data [5, 10] written to b1 is also written b2 since they are connected.

The Busses connected to each other can be disconnected. When disconnected, the data written to master is not written to slaves

julia> disconnect(b1, b2)

julia> isconnected(b1, b2)
false

Data Flow through Busses

Data flow through the Busses is very similar to the case in Links. See Data Flow through Links for information about data flow through Links. Runnable tasks must be bound to the links of the busses for data flow through the Bus. Again, put! and take! functions are used to write data from a Bus and read from data from a Bus.

Base.put! — Method

put!(bus::Bus, vals)

Puts vals to bus. Each item in vals is putted to the links of the bus.

Warning

The bus must be writable to be read. That is, there must be a runnable tasks bound to links of the bus that reads data from bus.

Base.take! — Method

take!(bus::Bus)

Takes an element from bus. Each link of the bus is a read and a vector containing the results is returned.

Warning

The bus must be readable to be read. That is, there must be a runnable tasks bound to links of the bus that writes data to bus.

Any data written to a Bus is recorded into the buffers of its links.

@repl writing_to_busses
using Jusdl # hide
b = Bus(2, 5);
launch(b);
put!(b, 1.);
b[1].buffer.data

Indexing and Iteration of Busses

Busses can be indexed similarly to the arrays in Julia. When indexed, corresponding link of the bus is returned.

julia> using Jusdl # hide

julia> b = Bus(3)
3-element Bus{Link{Float64}}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[1]
Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[end]
Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[:]
3-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[1] = Link()
Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> b[1:2] = [Link(), Link()]
2-element Array{Link{Float64},1}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

The iteration of Busses in a loop is also possible. When iterated, the links of the Bus is returned.

julia> using Jusdl # hide

julia> bus = Bus(3)
3-element Bus{Link{Float64}}:
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
 Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> for link in bus
           @show link
       end
link = Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
link = Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)
link = Link(state:open, eltype:Float64, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

Full API

Base.eltype — Function

eltype(link::Link)

Returns element type of link.

Missing docstring.

Missing docstring for Connections.length. Check Documenter's build log for details.

Base.iterate — Function

iterate(clk::Clock[, t=clk.t)

Iterationk interface for clk. clk can be iterated in a loop.

Example

julia> clk = Clock(0., 0.1, 0.3);

julia> set!(clk)
Clock(t:0.0, dt:0.1, tf:0.3, paused:false, isrunning:true)

julia> for t in clk
       @show t 
       end
t = 0.0
t = 0.1
t = 0.2
t = 0.3

Jusdl.Connections.hasslaves — Method

hasslaves(bus:Bus)

Returns true is all the links of bus has slaves. See also hasslaves(link::Link)

Jusdl.Connections.hasmaster — Method

hasmaster(bus::Bus)

Returns true is all the links of bus has master. See alsos hasmaster(link::Link)

Base.close — Method

close(bus::Bus)

Closes bus. When closed, no more data flow is possible for bus.

Jusdl.Connections.isfull — Method

isfull(bus::Bus)

Returns true when the links of bus are full.

Base.isreadable — Method

isreadable(bus::Bus)

Returns true if all the links of bus is readable.

Base.iswritable — Method

iswritable(bus::Bus)

Returns true if all the links of bus is writable.

Jusdl.Connections.snapshot — Method

snapshot(bus::Bus)

Returns all the data in links of bus.

Jusdl.Connections.launch — Method

launch(bus::Bus)

Launches every link of bus. See launch(link::Link)

Jusdl.Connections.launch — Method

launch(bus::Bus, valrange::AbstractVector)

Launches every links of bus with every item of valrange. See [launch(link:Link, valrange)(@ref)]

Base.getindex — Function

getindex(buf::Buffer, idx::Vararg{Int, N})

Returns an element from buf at index idx. Same as buf[idx] ```

getindex(bus::Bus, idx::Vararg{Int, N}) where N

Returns elements from bus at index idx.

Base.setindex! — Function

setindex!(buf::Buffer, val, idx)

Sets val to buf at index idx. Same as buf[idx] = val.

setindex!(bus::Bus, item, idx::Vararg{Int, N}) where N

Sets item to bus at index idx.