Links

Links are built on top of Channels of Julia. They are used as communication primitives for Tasks of Julia. A Link basically includes a Channel and a Buffer. The mode of the buffer is Cyclic.(see Buffer Modes for information on buffer modes). Every item sent through a Link is sent through the channel of the Link and written to the Buffer so that all the data flowing through a Link is recorded. The data transmitted through a Link can be of any Julia type, even if user-defined types.

Construction of Links

The construction of a Link is very simple: just specify its buffer length and element type.

Jusdl.Connections.Link — Type

Link{T}([ln::Int=64]) where T

Constructs a Link with element type T and buffer length ln. The buffer element type of T and mode is Cyclic.

Here, are some examples.

julia> using Jusdl # hide

julia> l1 = Link{Int}(5)
Link(state:open, eltype:Union{Missing, Int64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> l2 = Link{Matrix{Float64}}(10)
Link(state:open, eltype:Union{Missing, Array{Float64,2}}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

Similar to the case of Buffers, the data type that can flow the Link can be any Julia type, even a user-defined type.

julia> struct Object
           x::Int
       end

julia> l = Link{Object}(3)     # A `Link` that with element type `Object` with buffer size `3`.
Link(state:open, eltype:Union{Missing, Main.##ex-#387.Object}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

Connection and Disconnection of Links

Links can be connected to each other so that data can flow from one link to another. The flows from link l1 to l2, then l1 is said to drive l2 and l1 is called as master and l2 is called as slave. A Link can have more than one slave but can have just one master. When a Links is initialized, it has no master and slaves.

Jusdl.Connections.connect — Function

connect(master::Link, slave::Link)

Connects master to slave. When connected, the flow is from master to slave.

connect(links...)

Connect each link of links in the form of a path.

Example

julia> l1, l2 = Link(), Link();

julia> isconnected(l1, l2)
false

julia> connect(l1, l2)

julia> isconnected(l1, l2)
true

julia> ls = [Link() for i = 1 : 3];

julia> map(i -> isconnected(ls[i], ls[i + 1]), 1 : 2)
2-element Array{Bool,1}:
 0
 0

julia> connect(ls...)

julia> map(i -> isconnected(ls[i], ls[i + 1]), 1 : 2)
2-element Array{Bool,1}:
 1
 1

Similarly Links can be disconnected.

Jusdl.Connections.disconnect — Function

disconnect(link1::Link, link2::Link)

Disconnects link1 and link2. The order of arguments is not important.

Example

julia> ls = [Link() for i = 1 : 2];

julia> connect(ls[1], ls[2])

julia> disconnect(ls[1], ls[2])

julia> isconnected(ls[1], ls[2])
false

Warning

Note that the order or arguments is important when the links are connected. connect(l1, l2) connects l1 and l2 such that l1 drives l2, i.e., data flows from l1 to l2. In other words, l1 is the master link and l2 is the slave link. However, the order of arguments is not important when the links are disconnected. disconnect(l1, l2) does the same thing with disconnect(l2, l1), i.e., it justs breaks the connection between l2 and l1.

Data Flow through Links

The data can be read from and written into Links if active tasks are bound to them. Links can be thought of a pipe. In order to write data to a Link from one of its ends, a task that reads written data from the other end must be bounded to the Link. Similarly, in order to read data from one of the Link from one of its end, a task that writes the read data must be bound to the Link. Reading from and writing to Link is carried out with take! and put! functions. For more clarity, let us see some example.

Let us first construct a Link,

julia> l = Link{Float64}(5)
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

l is a Link with a buffer length of 5 and element type of Float64. Not that the l is open, but it is not ready for data reading or writing. To write data, we must bound a task that reads the written data.

julia> function reader(link::Link)  # Define job.
           while true
               val = take!(link)
               val === missing && break  # Poison-pill the tasks to terminate safely.
           end
       end
reader (generic function with 1 method)

julia> t = @async reader(l)
Task (runnable) @0x00007f920d2c09d0

The reader is defined such that the data written from one end of l is read until the data is missing. Now, we have runnable task t. This means the l is ready for data writing.

julia> put!(l, 1.)
1.0

julia> put!(l, 2.)
2.0

To terminate the task, we must write missing to l.

julia> put!(l, missing)  # Terminate the task
missing

julia> t   # Show that the `t` is terminated.
Task (done) @0x00007f920d2c09d0

Note that the data flown through the l is written to its buffer.

julia> l.buffer.data
5-element Array{Union{Missing, Float64},1}:
 1.0
 2.0
  missing
  missing
  missing

When ever the bound task to the l is runnable, the data can be written to l. That is, the data length that can be written to l is not limited by the buffer length of l. But, beware that the buffer of Linkss are Cyclic. That means, when the buffer is full, its data is overwritten.

julia> l = Link(5)
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> t = @async reader(l)
Task (runnable) @0x00007f920d77e980

julia> for item in 1. : 10.
           put!(l, item)
           @show l.buffer.data
       end
l.buffer.data = Union{Missing, Float64}[1.0, missing, missing, missing, missing]
l.buffer.data = Union{Missing, Float64}[1.0, 2.0, missing, missing, missing]
l.buffer.data = Union{Missing, Float64}[1.0, 2.0, 3.0, missing, missing]
l.buffer.data = Union{Missing, Float64}[1.0, 2.0, 3.0, 4.0, missing]
l.buffer.data = Union{Missing, Float64}[1.0, 2.0, 3.0, 4.0, 5.0]
l.buffer.data = Union{Missing, Float64}[6.0, 2.0, 3.0, 4.0, 5.0]
l.buffer.data = Union{Missing, Float64}[6.0, 7.0, 3.0, 4.0, 5.0]
l.buffer.data = Union{Missing, Float64}[6.0, 7.0, 8.0, 4.0, 5.0]
l.buffer.data = Union{Missing, Float64}[6.0, 7.0, 8.0, 9.0, 5.0]
l.buffer.data = Union{Missing, Float64}[6.0, 7.0, 8.0, 9.0, 10.0]

The case is very similar to read data from l. Again a runnable task is bound the l

julia> l = Link(5)
Link(state:open, eltype:Union{Missing, Float64}, hasmaster:false, numslaves:0, isreadable:false, iswritable:false)

julia> function writer(link::Link, vals)
           for val in vals
               put!(link, val)
           end
       end
writer (generic function with 1 method)

julia> t = @async writer(l, 1.:5.)
Task (runnable) @0x00007f920caa9fc0

julia> bind(l, t)
Channel{Union{Missing, Float64}}(sz_max:0,sz_curr:1)

julia> take!(l)
1.0

julia> take!(l)
2.0

It is possible to read data from l until t is active. To read all the data at once, collect can be used.

julia> t
Task (runnable) @0x00007f920caa9fc0

julia> collect(l)
3-element Array{Union{Missing, Float64},1}:
 3.0
 4.0
 5.0

julia> t  # Show that `t` is terminated.
Task (done) @0x00007f920caa9fc0

Full API

Base.put! — Function

put!(link::Link, val)

Puts val to link. val is handed over to the channel of link. val is also written in to the buffer of link.

Warning

link must be writable to put val. See launch

Example

julia> l = Link();

julia> t = launch(l);  # To be able to put values, `l` must be bound to a runnable task.

julia> put!(l, 1.)
┌ Info: Took 
└   val = 1.0
1.0

put!(bus::Bus, vals)

Puts vals to bus. Each item in vals is putted to the links of the bus.

Warning

The bus must be writable to be read. That is, there must be a runnable tasks bound to links of the bus that reads data from bus. See launch

Example

julia> b = Bus(2);

julia> t = launch(b);

julia> put!(b, [1., 2.])
┌ Info: Took 
└   val = 1.0
┌ Info: Took 
└   val = 2.0
2-element Array{Float64,1}:
 1.0
 2.0

Base.take! — Function

take!(link::Link)

Take an element from link.

Warning

link must be readable to take value. See launch

Example

julia> l = Link(5);

julia> t = launch(l, 1. : 5.);

julia> for i in 1 : 5
       @show take!(l)
       end
take!(l) = 1.0
take!(l) = 2.0
take!(l) = 3.0
take!(l) = 4.0
take!(l) = 5.0

take!(bus::Bus)

Takes an element from bus. Each link of the bus is a read and a vector containing the results is returned.

Warning

The bus must be readable to be read. That is, there must be a runnable tasks bound to links of the bus that writes data to bus. See launch

Example

julia> b = Bus(2);

julia> t = launch(b, [[rand() for i = 1 : 5] for j = 1 : 2])
2-element Array{Task,1}:
 Task (runnable) @0x00007f9634734280
 Task (runnable) @0x00007f96347344f0

julia> take!(b)
2-element Array{Float64,1}:
 0.6216364091492494
 0.0781964275368685

Base.close — Function

close(link)

Closes link. All the task bound the link is also terminated safely. When closed, data cannot ben into or read from the link.

Example

julia> l  = Link();

julia> t = launch(l);

julia> put!(l, 1.)
┌ Info: Took 
└   val = 1.0
1.0

julia> close(l)

julia> put!(l, 1.)
ERROR: InvalidStateException("Channel is closed.", :closed)
Stacktrace:
 [1] check_channel_state at ./channels.jl:167 [inlined]
 [2] put! at ./channels.jl:323 [inlined]
 [3] put!(::Link{Union{Missing, Float64}}, ::Float64) at /home/sari/.julia/dev/Jusdl/src/connections/link.jl:64
 [4] top-level scope at REPL[99]:1

close(bus::Bus)

Closes bus. When closed, no more data flow is possible for bus.

Base.isopen — Function

open(link::Link)

Returns true if link is open. A link is open if its channel is open.

Base.isreadable — Function

isreadable(link::Link)

Returns true if link is readable. When link is readable, data can be read from link with take function.

isreadable(bus::Bus)

Returns true if all the links of bus is readable.

Base.iswritable — Function

writable(link::Link)

Returns true if link is writable. When link is writable, data can be written into link with put function.

iswritable(bus::Bus)

Returns true if all the links of bus is writable.

Jusdl.Connections.isfull — Function

isfull(link::Link)

Returns true if the buffer of link is full.

isfull(bus::Bus)

Returns true when the links of bus are full.

Jusdl.Connections.isconnected — Function

isconnected(link1, link2)

Returns true if link1 is connected to link2. The order of the arguments are not important.

Jusdl.Connections.hasslaves — Function

hasslaves(link::Link)

Returns true if link has slave links.

Jusdl.Connections.hasmaster — Function

hasmaster(link::Link)

Returns true if link has a master link.

Jusdl.Connections.getmaster — Function

getmaster(link::Link)

Returns the master of link.

Jusdl.Connections.getslaves — Function

getslaves(link::Link)

Returns the slaves of link.

Jusdl.Connections.snapshot — Function

snapshot(link::Link)

Returns all the data of the buffer of link.

snapshot(bus::Bus)

Returns all the data in links of bus.

Jusdl.Connections.UnconnectedLinkError — Type

UnconnectedLinkError <: Exception

Exception thrown when the links are not connected to each other.

Jusdl.Connections.Pin — Type

Pin()

Constructs a Pin. A Pin is the auxilary type to monitor connection status of Links. See Link

Jusdl.Connections.findflow — Function

findflow(link1::Link, link2::Link)

Returns a tuple of (masterlink, slavelink) where masterlink is the link that drives the other and slavelink is the link that is driven by the other.

Example

julia> ls = [Link() for i = 1 : 2];

julia> connect(ls[1], ls[2])

julia> findflow(ls[2], ls[1]) .== (ls[1], ls[2])
(true, true)

Jusdl.Connections.insert — Function

insert(master::Link, slave::Link, new::Link)

Inserts the new link between the master link and slave link. The master is connected to new, and new is connected to slave.

Example

julia> ls = [Link() for i = 1 : 3];  

julia> connect(ls[1], ls[2]) 

julia> insert(ls[1], ls[2], ls[3])

julia> isconnected(ls[1], ls[2])
false

julia> isconnected(ls[1], ls[3]) && isconnected(ls[3], ls[2])
true

Jusdl.Connections.release — Function

release(link::Link)

Release all the slave links of link. That is, all the slave links of link is disconnected.

Example

julia> ls = [Link() for i = 1 : 5];

julia> foreach(l -> connect(ls[1], l), ls[2:5])

julia> map(l -> isconnected(ls[1], l), ls[2:5])
4-element Array{Bool,1}:
 1
 1
 1
 1

julia> release(ls[1])  # Release all the slaves.

julia> map(l -> isconnected(ls[1], l), ls[2:5])
4-element Array{Bool,1}:
 0
 0
 0
 0

Base.bind — Function

bind(link::Link, task::Task)

Binds task to link. When task is done link is closed.

Base.collect — Function

collect(link::Link)

Collects all the available data on the link.

Jusdl.Connections.launch — Function

launch(link::Link)

Constructs a taker task and binds it to link. The taker task reads the data and prints an info message until missing is read from the link.

launch(link:Link, valrange)

Constructs a putter task and binds it to link. putter tasks puts the data in valrange.

launch(bus::Bus)

Launches every link of bus. See launch(link::Link)

launch(bus::Bus, valrange::AbstractVector)

Launches every links of bus with every item of valrange. See [launch(link:Link, valrange)(@ref)]