Buffer

Buffer is a primitive to buffer the data. Data can be of any Julia type. Data can be read from and written into a buffer, and the mode of the buffer determines the way to read from and write into the buffers.

Buffer Modes

Buffer mode determines the way the data is read from and written into a Buffer.

Jusdl.Utilities.BufferMode — Type

BufferMode

Abstract type for buffer mode. Subtypes of BufferMode is CyclicMode and LinearMode.

Jusdl.Utilities.LinearMode — Type

LinearMode <: BufferMode

Abstract type of linear buffer modes. See Normal, Lifo, Fifo

Jusdl.Utilities.CyclicMode — Type

CyclicMode <: BufferMode

Abstract type of cyclic buffer modes. See Cyclic

There are four different buffer modes.

Jusdl.Utilities.Normal — Type

Normal <: LinearMode

LinearMode buffer mode. The data is written to buffer until the buffer is full. When it is full, no more data is written to the buffer. When read, the data written last is returned and the returned data is not deleted from the internal container of the buffer.

Jusdl.Utilities.Cyclic — Type

Cyclic <: CyclicMode

Cyclic buffer mode. The data is written to buffer until the buffer is full. When the buffer is full, new data is written by overwriting the data available in the buffer starting from the beginning of the buffer. When the buffer is read, the element written last is returned and the returned element is not deleted from the buffer.

Jusdl.Utilities.Lifo — Type

Lifo <: LinearMode

Lifo (Last-in-first-out) buffer mode. This type of buffer is a last-in-first-out buffer. Data is written to the buffer until the buffer is full. When the buffer is full, no more element can be written into the buffer. When read, the last element written into buffer is returned. The returned element is deleted from the buffer.

Jusdl.Utilities.Fifo — Type

Fifo <: LinearMode

Fifo (First-in-last-out) buffer mode. This type of buffer is a first-in-first-out buffer. The data is written to the buffer until the buffer is full. When the buffer is full, no more element can be written into the buffer. When read, the first element written into the buffer is returned. The returned element is deleted from the buffer.

Buffer Constructors

The Buffer construction is very similar to the construction of arrays in Julia. Just specify the mode, element type and length of the buffer. Here are the main Buffer constructors:

Jusdl.Utilities.Buffer — Type

Buffer{M}(::Type{T}, ln::Int) where {M, T}

Constructs a Buffer of length ln with element type of T. M is the mode of the Buffer that determines how data is to read from and written into the Buffer. There exists for different buffer modes:

Normal: See Normal

Cyclic: See Cyclic

Lifo: See Lifo

Fifo: See Fifo

The default mode for Buffer is Cyclic and default element type is Float64.

Buffer(::Type{T}, ln::Int) where T

Constructs a Buffer of length ln and with element type of T. The mode of the buffer is Cyclic.

Buffer{M}(ln::Int) where M

Constructs a Buffer of length of ln and with mode M. M can be Normal, Cyclic, Fifo and Lifo. The element type of the Buffer is Float64.

Buffer(ln::Int)

Constructs a Buffer of length ln with mode Cyclic and element type of Float64.

Warning

Note that Buffer is one dimensional. That is, the length of the data must be specified when constructing a Buffer.

Warning

Note that when a Buffer is initialized, the internal data of the Buffer is of missing.

Let us try some examples. Here are some simple buffer construction.

julia> buf1 = Buffer{Normal}(Float64, 5)   # Buffer of length `5` with mode `Normal` and element type of `Float64`.
Buffer(mode:Normal, eltype:Union{Missing, Float64}, length:5, index:1, state:empty)

julia> buf2 = Buffer{Fifo}(Int, 3)       # Buffer of length `5` with mode `Fifo` and element type of `Int`.
Buffer(mode:Fifo, eltype:Union{Missing, Int64}, length:3, index:1, state:empty)

julia> buf3 = Buffer(Vector{Int}, 3)       # Buffer of length `5` with mode `Cyclic` and element type of `Vector{Int}`.
Buffer(mode:Cyclic, eltype:Union{Missing, Array{Int64,1}}, length:3, index:1, state:empty)

julia> buf4 = Buffer(Matrix{Float64}, 5)    # Buffer of length `5` with mode `Cyclic` and element type of `Matrix{Float64}`.
Buffer(mode:Cyclic, eltype:Union{Missing, Array{Float64,2}}, length:5, index:1, state:empty)

julia> buf5 = Buffer(5)                    # Buffer of length `5` with mode `Cyclic` and element type of `Float64`.
Buffer(mode:Cyclic, eltype:Union{Missing, Float64}, length:5, index:1, state:empty)

Note that the element type of Buffer can be any Julia type, even any user-defined type. Note the following example,

julia> using Jusdl #hide

julia> struct Object end       # Define a dummy type.

julia> buf = Buffer{Normal}(Object, 4)  # Buffer of length `4` with element type `Object`.
Buffer(mode:Normal, eltype:Union{Main.##ex-#423.Object, Missing}, length:4, index:1, state:empty)

Writing Data into Buffers

Writing data into a Buffer is done with write! function.

Jusdl.Utilities.write! — Function

write!(buf::Buffer{M, T}, val) where {M, T}

Writes val into buf. Writing is carried occurding the mode M of buf. See Normal, Cyclic, Lifo, Fifo for buffer modes.

Example

julia> buf = Buffer(3)
Buffer(mode:Cyclic, eltype:Union{Missing, Float64}, length:3, index:1, state:empty)

julia> buf.data  # Initailly all the elements of `buf` is missing.
3-element Array{Union{Missing, Float64},1}:
 missing
 missing
 missing

julia> write!(buf, 3.)
3.0

julia> buf.data
3-element Array{Union{Missing, Float64},1}:
 3.0     
  missing
  missing

Recall that when the buffer is full, no more data can be written into the buffer if the buffer mode is of type LinearMode.

julia> normalbuf = Buffer{Normal}(3)
Buffer(mode:Normal, eltype:Union{Missing, Float64}, length:3, index:1, state:empty)

julia> fill!(normalbuf, 1.)

julia> normalbuf.data
3-element Array{Union{Missing, Float64},1}:
 1.0
 1.0
 1.0

julia> write!(normalbuf, 1.)
┌ Warning: Buffer is full.
└ @ Jusdl.Utilities ~/.julia/dev/Jusdl/src/utilities/buffers.jl:153
1.0

This situation is the same for Lifo and Fifo buffers, but not the case for Cyclic buffer.

julia> normalbuf = Buffer{Cyclic}(3)
Buffer(mode:Cyclic, eltype:Union{Missing, Float64}, length:3, index:1, state:empty)

julia> fill!(normalbuf, 1.)

julia> normalbuf.data
3-element Array{Union{Missing, Float64},1}:
 1.0
 1.0
 1.0

julia> write!(normalbuf, 3.)
3.0

julia> write!(normalbuf, 4.)
4.0

Warning

Since when a Buffer is constructed, it is empty, no data is written to it. But it is initialized with missing data. Thus, the element type of buffer of type Buffer{M, T} where {M, T} is Union{Missing, T} where T. Benchmarks that has been carried out shows that there is no performance bottle neck is such design since Julia's compiler can compile optimized code for such a small unions. Therefore it is possible to write missing into a buffer of type Buffer{M,T} where {M,T}.

Reading Data from Buffers

Reading data from a Buffer is done with read function.

Base.read — Function

read(buf::Buffer)

Reads an element from buf. Reading is performed according to the mode of buf. See Normal, Cyclic, Lifo, Fifo for buffer modes.

Example

julia> buf = Buffer{Fifo}(3)
Buffer(mode:Fifo, eltype:Union{Missing, Float64}, length:3, index:1, state:empty)

julia> for val in 1 : 3. 
       write!(buf, val)
       @show buf.data
       end 
buf.data = Union{Missing, Float64}[1.0, missing, missing]
buf.data = Union{Missing, Float64}[1.0, 2.0, missing]
buf.data = Union{Missing, Float64}[1.0, 2.0, 3.0]

julia> for i in 1 : 3 
       item = read(buf)
       @show (item, buf.data)
       end
(item, buf.data) = (1.0, Union{Missing, Float64}[2.0, 3.0, missing])
(item, buf.data) = (2.0, Union{Missing, Float64}[3.0, missing, missing])
(item, buf.data) = (3.0, Union{Missing, Float64}[missing, missing, missing])

AbstractArray Interface of Buffers

A Buffer can be indexed using the similar syntax of arrays in Julia. That is, getindex and setindex! methods can be used with known Julia syntax. i.e. getindex(buf, idx) is equal to buf[idx] and setindex(buf, val, idx) is equal to buf[idx] = val.

julia> buf = Buffer(5)
Buffer(mode:Cyclic, eltype:Union{Missing, Float64}, length:5, index:1, state:empty)

julia> size(buf)
(5,)

julia> length(buf)
5

julia> for val in 1 : 5
           write!(buf, 2val)
       end

julia> buf[1]
2.0

julia> buf[3:4]
2-element Array{Union{Missing, Float64},1}:
 6.0
 8.0

julia> buf[[3, 5]]
2-element Array{Union{Missing, Float64},1}:
  6.0
 10.0

julia> buf[end]
10.0

julia> buf[1] = 5
5

julia> buf[3:5] = [7, 8, 9]
3-element Array{Int64,1}:
 7
 8
 9

Full API

Base.fill! — Function

fill!(buf::Buffer{M, T}, val::T) where {M,T}

Writes val into buf until buf is full.

Base.isempty — Function

isempty(buf::Buffer)

Returns true if buf is empty.

Jusdl.Utilities.isfull — Function

isfull(buf::Buffer)

Returns true if buf is full.

Jusdl.Utilities.content — Function

content(buf, [flip=true])

Returns the current data of buf. If flip is true, the data to be returned is flipped.